+++ data/dataset.py
@@ -1,24 +1,26 @@
+import random
+import numpy as np
+from PIL import Image
+
 import torch
 from torch.utils.data import Dataset
-import numpy as np
 from torchvision import transforms
-import random
-from PIL import Image
 import torchvision.transforms as T
+
 
 class TrackableRandomHorizontalFlip(T.RandomHorizontalFlip):
     def __init__(self, p=0.5):
         super().__init__(p)
-        self.flipped = False  # Initialize a flag to track the flip
+        self.flipped = False
     
     def __call__(self, img):
-        # Decide whether to flip
         if random.random() < self.p:
             self.flipped = True
             return img.transpose(Image.FLIP_LEFT_RIGHT)
         else:
             self.flipped = False
             return img
+
 
 class WiderFaceDataset(Dataset):
     def __init__(self, data, labels, train=True):
@@ -31,100 +33,81 @@ class WiderFaceDataset(Dataset):
         self.data = data
         self.labels = labels
         self.train = train
-        ## read in coordinate here as a key value map between filename and coords
-
-        ## read in filename as a list
 
     def __len__(self):
-        # Return the number of samples
         return len(self.data)
 
     def __getitem__(self, idx):
-        # Get data and label for a given index
         img_file = self.data[idx]
+        
         if self.train:
-            img = Image.open("YOCO3k/train/images/" + img_file)
+            o_img = Image.open("datasets/YOCO3k/train/images/" + img_file)
         else:
-            img = Image.open("WIDER_val/images/" + img_file)
+            o_img = Image.open("datasets/YOCO3k/val/images/" + img_file)
         label = self.labels[idx]
 
-        # raw coords, label: [[xmin, ymin, width, height], ...] (seq_len, 4)
-        # coords [[y, x], [y, x]...]
         coords = np.zeros((label.shape[0], 2))
-        if len(coords) > 0:
-            coords[:, 1] = label[:, 0] + (label[:, 2]-1)/2
-            coords[:, 0] = label[:, 1] + (label[:, 3]-1)/2
-            img, coords = self.resize_and_transform_coord(img, coords, self.train)
-        else:
-            img, coords = self.resize(img, self.train)
-        # sample: (batch, 3, H, W)
-        # coordinates: [batch, seq_len, 2]
-        # heatmaps: ? [batch, seq_len, 1, H, W] (TYPICALLY GENERATED ON THE FLY)
-        return img, coords
 
-    # resize the img to 384x384
-    # transform original coord (in list?)
-    # coords will be used for finding adaptive arrangement 
-    def resize_and_transform_coord(self, img, coords, train):
+        coords[:, 1] = label[:, 0] + (label[:, 2]-1)/2
+        coords[:, 0] = label[:, 1] + (label[:, 3]-1)/2
+
+        label[:,2] += (label[:, 0]-1)
+        label[:,3] += (label[:, 1]-1)
+        label = label[:, [0, 2, 1, 3]] #(xmin, xmax, ymin, ymax)
+
+        img, coords, bbox = self.resize_and_transform_coord(o_img, coords, label, self.train)
+
+        return o_img, img, coords, bbox
+
+    def resize_and_transform_coord(self, img, coords, label, train):
         width, height = img.size
         h_ratio = 384 / height
         w_ratio = 384 / width
+
         if train:
             transform = transforms.Compose([
                     transforms.Resize((384, 384), interpolation=transforms.InterpolationMode.LANCZOS),
                     TrackableRandomHorizontalFlip(p=0.5),
-                    transforms.ToTensor(),  # Convert to tensor
+                    transforms.ToTensor(),
                     transforms.Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5]),
                 ])
             img = transform(img)
             flipped = transform.transforms[1].flipped
             if flipped:
                 coords[:,1] = width - coords[:, 1] - 1
-        else:
-            transform = transforms.Compose([
-                    transforms.Resize((384, 384), interpolation=transforms.InterpolationMode.LANCZOS),
-                    transforms.ToTensor(),  # Convert to tensor
-                    transforms.Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5]),
-                ])
-            img = transform(img)
-        coords[:,0] = np.clip(coords[:,0] * h_ratio, 0, 383) # y coord
-        coords[:,1] = np.clip(coords[:,1] * w_ratio, 0, 383) # x coord
-        return img, coords
+                label[:,0] = width - label[:,0] - 1
+                label[:,2] = width - label[:,2] - 1
 
-    def resize(self, img, train):
-        if train:
-            transform = transforms.Compose([
-                    transforms.Resize((384, 384), interpolation=transforms.InterpolationMode.LANCZOS),
-                    TrackableRandomHorizontalFlip(p=0.5),
-                    transforms.ToTensor(),  # Convert to tensor
-                    transforms.Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5]),
-                ])
-            img = transform(img)
         else:
             transform = transforms.Compose([
                     transforms.Resize((384, 384), interpolation=transforms.InterpolationMode.LANCZOS),
-                    transforms.ToTensor(),  # Convert to tensor
+                    transforms.ToTensor(),
                     transforms.Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5]),
                 ])
             img = transform(img)
         
-        return img, np.array([])
+        coords[:,0] = np.clip(coords[:,0] * h_ratio, 0, 383) # y coord
+        coords[:,1] = np.clip(coords[:,1] * w_ratio, 0, 383) # x coord
+
+        label[:,:2] = label[:, :2] * w_ratio
+        label[:,2:4] = label[:,2:4] * h_ratio
+
+        return img, coords, label
 
     def custom_collate_fn(self, batch):
-        images = torch.stack([item[0] for item in batch])  # Images
-        seq_lens = torch.tensor([len(item[1]) for item in batch], dtype=torch.long)
-        max_seq_len = torch.max(seq_lens) + 1 # the termination step
-        coordinates = torch.ones((len(batch), max_seq_len, 2), dtype=torch.float) * 10000  # make other ground truth far away
-        #coordinates = [item[1] for item in batch]  # List of variable-length tensors
+        o_images = [item[0] for item in batch]
+        images = torch.stack([item[1] for item in batch])
+
+        seq_lens = torch.tensor([len(item[2]) for item in batch], dtype=torch.long)
+        max_seq_len = torch.max(seq_lens) + 1
+
+        coordinates = torch.ones((len(batch), max_seq_len, 2), dtype=torch.float) * 10000
+        bboxes = torch.ones((len(batch), max_seq_len, 4), dtype=torch.float) * 10000
+
         for i, item in enumerate(batch):
-            seq_len = seq_lens[i]  # Current sequence length
-            if len(item[1]) > 0:
-                coordinates[i, :seq_len, :] = torch.tensor(item[1], dtype=torch.float)  # Copy label into coordinates
-        return images, coordinates, seq_lens, max_seq_len
+            seq_len = seq_lens[i]
+            if len(item[2]) > 0:
+                coordinates[i, :seq_len, :] = torch.tensor(item[2], dtype=torch.float)
+                bboxes[i, :seq_len, :] = torch.tensor(item[3], dtype=torch.float)
 
-
-
-    
-
-
-
+        return o_images, images, coordinates, seq_lens, max_seq_len, bboxes
