+++ environment.yaml
@@ -1,0 +1,89 @@
+name: yoco
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - _openmp_mutex=5.1=1_gnu
+  - ca-certificates=2024.11.26=h06a4308_0
+  - ld_impl_linux-64=2.40=h12ee557_0
+  - libffi=3.4.4=h6a678d5_1
+  - libgcc-ng=11.2.0=h1234567_1
+  - libgomp=11.2.0=h1234567_1
+  - libstdcxx-ng=11.2.0=h1234567_1
+  - ncurses=6.4=h6a678d5_0
+  - openssl=3.0.15=h5eee18b_0
+  - pip=24.2=py39h06a4308_0
+  - python=3.9.20=he870216_1
+  - readline=8.2=h5eee18b_0
+  - setuptools=75.1.0=py39h06a4308_0
+  - sqlite=3.45.3=h5eee18b_0
+  - tk=8.6.14=h39e8969_0
+  - tzdata=2024b=h04d1e81_0
+  - wheel=0.44.0=py39h06a4308_0
+  - xz=5.4.6=h5eee18b_1
+  - zlib=1.2.13=h5eee18b_1
+  - pip:
+      - antlr4-python3-runtime==4.9.3
+      - certifi==2024.8.30
+      - charset-normalizer==3.4.0
+      - click==8.1.7
+      - contourpy==1.3.0
+      - cycler==0.12.1
+      - docker-pycreds==0.4.0
+      - filelock==3.13.1
+      - fonttools==4.55.0
+      - fsspec==2024.2.0
+      - gitdb==4.0.11
+      - gitpython==3.1.43
+      - huggingface-hub==0.26.3
+      - idna==3.10
+      - importlib-resources==6.4.5
+      - jinja2==3.1.3
+      - kiwisolver==1.4.7
+      - markupsafe==2.1.5
+      - matplotlib==3.9.3
+      - mpmath==1.3.0
+      - networkx==3.2.1
+      - numpy==1.26.3
+      - nvidia-cublas-cu12==12.1.3.1
+      - nvidia-cuda-cupti-cu12==12.1.105
+      - nvidia-cuda-nvrtc-cu12==12.1.105
+      - nvidia-cuda-runtime-cu12==12.1.105
+      - nvidia-cudnn-cu12==8.9.2.26
+      - nvidia-cufft-cu12==11.0.2.54
+      - nvidia-curand-cu12==10.3.2.106
+      - nvidia-cusolver-cu12==11.4.5.107
+      - nvidia-cusparse-cu12==12.1.0.106
+      - nvidia-nccl-cu12==2.20.5
+      - nvidia-nvjitlink-cu12==12.1.105
+      - nvidia-nvtx-cu12==12.1.105
+      - omegaconf==2.3.0
+      - packaging==24.2
+      - pillow==10.2.0
+      - platformdirs==4.3.6
+      - protobuf==5.29.0
+      - psutil==6.1.0
+      - pyparsing==3.2.0
+      - python-dateutil==2.9.0.post0
+      - pyyaml==6.0.2
+      - regex==2024.11.6
+      - requests==2.32.3
+      - safetensors==0.4.5
+      - scipy==1.13.1
+      - sentry-sdk==2.19.0
+      - setproctitle==1.3.4
+      - six==1.16.0
+      - smmap==5.0.1
+      - sympy==1.13.1
+      - tokenizers==0.20.3
+      - torch==2.3.1+cu121
+      - torchaudio==2.3.1+cu121
+      - torchvision==0.18.1+cu121
+      - tqdm==4.67.1
+      - transformers==4.46.3
+      - triton==2.3.1
+      - typing-extensions==4.9.0
+      - urllib3==2.2.3
+      - wandb==0.18.7
+      - zipp==3.21.0
