--- generalize_data.py
@@ -1,2 +1,0 @@
-gen_data = ['2--Demonstration/2_Demonstration_Protesters_2_268.jpg', '3--Riot/3_Riot_Riot_3_306.jpg', '51--Dresses/51_Dresses_wearingdress_51_414.jpg', '37--Soccer/37_Soccer_soccer_ball_37_345.jpg', '7--Cheering/7_Cheering_Cheering_7_313.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_114.jpg', '12--Group/12_Group_Group_12_Group_Group_12_84.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_223.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_311.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_662.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_94.jpg', '13--Interview/13_Interview_Interview_Sequences_13_7.jpg', '6--Funeral/6_Funeral_Funeral_6_1029.jpg', '4--Dancing/4_Dancing_Dancing_4_57.jpg', '2--Demonstration/2_Demonstration_Protesters_2_796.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_196.jpg', '56--Voter/56_Voter_peoplevoting_56_663.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_493.jpg', '48--Parachutist_Paratrooper/48_Parachutist_Paratrooper_Parachutist_Paratrooper_48_678.jpg', '22--Picnic/22_Picnic_Picnic_22_541.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_71.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_736.jpg', '35--Basketball/35_Basketball_Basketball_35_180.jpg', '52--Photographers/52_Photographers_taketouristphotos_52_288.jpg', '2--Demonstration/2_Demonstration_Protesters_2_779.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_48.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_72.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_143.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_550.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_341.jpg', '0--Parade/0_Parade_Parade_0_854.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_206.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_665.jpg', '13--Interview/13_Interview_Interview_Sequences_13_495.jpg', '22--Picnic/22_Picnic_Picnic_22_36.jpg', '53--Raid/53_Raid_policeraid_53_207.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_419.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_914.jpg', '2--Demonstration/2_Demonstration_Protesters_2_352.jpg', '53--Raid/53_Raid_policeraid_53_368.jpg', '3--Riot/3_Riot_Riot_3_790.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_171.jpg', '51--Dresses/51_Dresses_wearingdress_51_7.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_637.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_759.jpg', '53--Raid/53_Raid_policeraid_53_445.jpg', '0--Parade/0_Parade_Parade_0_465.jpg', '28--Sports_Fan/28_Sports_Fan_Sports_Fan_28_118.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_90.jpg', '13--Interview/13_Interview_Interview_On_Location_13_521.jpg', '6--Funeral/6_Funeral_Funeral_6_128.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_17.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_138.jpg', '0--Parade/0_Parade_Parade_0_218.jpg', '6--Funeral/6_Funeral_Funeral_6_760.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_306.jpg', '21--Festival/21_Festival_Festival_21_526.jpg', '46--Jockey/46_Jockey_Jockey_46_79.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_424.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_107.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_510.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_339.jpg', '38--Tennis/38_Tennis_Tennis_38_580.jpg', '59--people--driving--car/59_peopledrivingcar_peopledrivingcar_59_532.jpg', '3--Riot/3_Riot_Riot_3_393.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_66.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_252.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_654.jpg', '43--Row_Boat/43_Row_Boat_Canoe_43_325.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_571.jpg', '0--Parade/0_Parade_Parade_0_478.jpg', '42--Car_Racing/42_Car_Racing_Car_Racing_42_906.jpg', '8--Election_Campain/8_Election_Campain_Election_Campaign_8_86.jpg', '35--Basketball/35_Basketball_Basketball_35_664.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_460.jpg', '2--Demonstration/2_Demonstration_Protesters_2_148.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_679.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_430.jpg', '0--Parade/0_Parade_marchingband_1_556.jpg', '21--Festival/21_Festival_Festival_21_976.jpg', '35--Basketball/35_Basketball_playingbasketball_35_65.jpg', '13--Interview/13_Interview_Interview_On_Location_13_401.jpg', '12--Group/12_Group_Group_12_Group_Group_12_519.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_619.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_700.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_216.jpg', '4--Dancing/4_Dancing_Dancing_4_41.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_1007.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_50.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_368.jpg', '45--Balloonist/45_Balloonist_Balloonist_45_273.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_60.jpg', '36--Football/36_Football_Football_36_110.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_601.jpg', '35--Basketball/35_Basketball_Basketball_35_653.jpg', '3--Riot/3_Riot_Riot_3_263.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_325.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_matadorbullfighting_47_207.jpg', '22--Picnic/22_Picnic_Picnic_22_308.jpg', '12--Group/12_Group_Group_12_Group_Group_12_354.jpg', '41--Swimming/41_Swimming_Swimming_41_243.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_653.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_326.jpg', '57--Angler/57_Angler_peoplefishing_57_661.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_43.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_932.jpg', '3--Riot/3_Riot_Riot_3_522.jpg', '46--Jockey/46_Jockey_Jockey_46_444.jpg', '8--Election_Campain/8_Election_Campain_Election_Campaign_8_297.jpg', '56--Voter/56_Voter_peoplevoting_56_344.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_223.jpg', '2--Demonstration/2_Demonstration_Protesters_2_583.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_171.jpg', '6--Funeral/6_Funeral_Funeral_6_618.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_615.jpg', '35--Basketball/35_Basketball_Basketball_35_712.jpg', '2--Demonstration/2_Demonstration_Protesters_2_901.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_643.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_414.jpg', '2--Demonstration/2_Demonstration_Protesters_2_884.jpg', '0--Parade/0_Parade_marchingband_1_309.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_349.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_31.jpg', '21--Festival/21_Festival_Festival_21_331.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_494.jpg', '53--Raid/53_Raid_policeraid_53_599.jpg', '53--Raid/53_Raid_policeraid_53_696.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_246.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_114.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_395.jpg', '0--Parade/0_Parade_marchingband_1_382.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_1001.jpg', '7--Cheering/7_Cheering_Cheering_7_138.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_176.jpg', '7--Cheering/7_Cheering_Cheering_7_391.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_18.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_43.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_404.jpg', '8--Election_Campain/8_Election_Campain_Election_Campaign_8_244.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_86.jpg']
-gen_label = [[[74, 204, 41, 56], [189, 208, 25, 58], [282, 243, 30, 40], [366, 239, 24, 36], [396, 232, 32, 40], [471, 230, 32, 52], [534, 257, 28, 33], [592, 206, 36, 44], [691, 233, 25, 35], [843, 209, 37, 45], [987, 257, 24, 26]], [[263, 142, 63, 65], [479, 191, 57, 60], [659, 169, 71, 71], [746, 81, 63, 85], [924, 67, 59, 79], [145, 151, 22, 23], [472, 134, 35, 54], [324, 130, 16, 25], [238, 189, 29, 43], [289, 235, 35, 72], [841, 130, 12, 17]], [[697, 97, 79, 101], [516, 22, 121, 227], [425, 63, 58, 71], [380, 48, 46, 67], [335, 120, 61, 80], [194, 99, 94, 110], [134, 0, 55, 40], [257, 0, 40, 26], [79, 120, 61, 76], [1, 137, 74, 97], [12, 70, 62, 81]], [[141, 315, 18, 27], [1001, 190, 15, 21], [969, 180, 16, 23], [818, 167, 15, 20], [715, 164, 15, 19], [384, 93, 11, 14], [365, 179, 15, 21], [427, 182, 15, 21], [555, 166, 16, 22], [490, 171, 16, 21], [622, 179, 15, 21]], [[5, 90, 21, 31], [29, 109, 32, 52], [161, 121, 50, 66], [232, 142, 41, 57], [343, 154, 39, 59], [468, 239, 34, 69], [652, 148, 35, 54], [746, 131, 40, 69], [951, 181, 39, 67], [972, 140, 34, 71], [842, 0, 21, 16]], [[914, 236, 30, 36], [908, 124, 23, 35], [943, 138, 22, 32], [825, 123, 24, 34], [811, 232, 24, 32], [769, 379, 47, 60], [745, 294, 24, 24], [587, 290, 71, 94], [499, 326, 27, 54], [364, 142, 81, 142], [123, 306, 84, 126]], [[119, 129, 49, 63], [281, 76, 45, 64], [331, 111, 49, 71], [432, 112, 53, 68], [524, 75, 48, 65], [369, 302, 57, 74], [178, 307, 55, 57], [577, 286, 60, 89], [606, 126, 50, 71], [823, 102, 62, 80], [795, 316, 78, 95]], [[122, 48, 40, 51], [278, 108, 32, 39], [419, 51, 43, 54], [570, 108, 34, 40], [696, 64, 39, 51], [822, 108, 31, 39], [120, 318, 41, 52], [292, 376, 31, 37], [455, 373, 31, 40], [652, 319, 38, 48], [865, 323, 38, 44]], [[684, 409, 22, 18], [829, 426, 10, 11], [447, 366, 12, 14], [446, 396, 14, 17], [300, 381, 14, 15], [296, 374, 11, 11], [234, 380, 18, 20], [203, 380, 14, 15], [47, 374, 15, 17], [120, 371, 12, 13], [0, 362, 9, 15]], [[397, 207, 20, 29], [265, 220, 18, 22], [231, 234, 12, 16], [552, 300, 30, 19], [445, 272, 25, 26], [802, 302, 44, 39], [986, 207, 38, 45], [28, 206, 13, 35], [446, 400, 15, 25], [356, 247, 10, 11], [119, 251, 5, 7]], [[820, 243, 42, 44], [689, 171, 41, 45], [762, 120, 39, 52], [600, 100, 34, 57], [438, 116, 39, 51], [569, 162, 44, 38], [303, 97, 45, 45], [1009, 108, 15, 51], [977, 112, 31, 46], [126, 41, 36, 39], [47, 12, 34, 51]], [[106, 247, 23, 29], [253, 206, 36, 47], [384, 265, 21, 23], [476, 249, 25, 30], [637, 192, 37, 53], [573, 305, 15, 14], [783, 278, 13, 23], [768, 296, 13, 17], [778, 316, 12, 14], [863, 244, 23, 26], [916, 255, 24, 27]], [[272, 123, 110, 138], [418, 206, 37, 59], [145, 164, 39, 42], [80, 146, 29, 44], [41, 176, 38, 50], [0, 201, 30, 35], [592, 114, 109, 129], [837, 103, 55, 67], [981, 100, 43, 78], [363, 169, 34, 62], [232, 208, 30, 54]], [[559, 137, 32, 53], [592, 199, 37, 69], [733, 157, 24, 34], [846, 160, 28, 44], [817, 162, 18, 31], [781, 190, 18, 29], [947, 199, 19, 26], [893, 163, 20, 29], [772, 168, 24, 32], [1005, 182, 19, 33], [969, 129, 20, 28]], [[407, 242, 153, 186], [359, 75, 29, 38], [352, 129, 29, 36], [487, 131, 34, 42], [390, 9, 36, 48], [522, 101, 30, 31], [562, 94, 26, 28], [601, 61, 26, 31], [578, 141, 29, 33], [608, 141, 39, 37], [673, 76, 61, 68]], [[875, 86, 49, 61], [805, 141, 41, 58], [615, 253, 35, 46], [542, 193, 35, 47], [475, 317, 32, 42], [404, 280, 33, 44], [309, 194, 37, 47], [232, 195, 36, 51], [164, 192, 41, 51], [102, 107, 53, 69], [716, 134, 45, 57]], [[670, 148, 42, 41], [489, 285, 51, 72], [342, 207, 51, 49], [106, 225, 12, 13], [2, 218, 8, 12], [196, 196, 13, 19], [71, 218, 11, 11], [282, 190, 11, 11], [233, 176, 10, 15], [316, 184, 9, 12], [839, 64, 62, 74]], [[18, 188, 45, 48], [222, 178, 33, 33], [248, 140, 42, 49], [220, 310, 52, 50], [379, 124, 48, 55], [566, 105, 40, 61], [619, 150, 48, 56], [754, 162, 46, 53], [877, 123, 29, 57], [972, 182, 37, 44], [586, 463, 23, 50]], [[489, 268, 18, 26], [555, 242, 15, 23], [580, 320, 16, 27], [603, 328, 19, 27], [623, 330, 22, 33], [651, 309, 52, 64], [335, 334, 43, 54], [397, 351, 27, 34], [427, 336, 21, 30], [461, 332, 21, 29], [253, 288, 74, 122]], [[90, 634, 42, 65], [360, 1109, 12, 19], [263, 1081, 5, 10], [269, 1092, 5, 9], [903, 846, 18, 17], [447, 1184, 3, 5], [454, 1188, 3, 4], [459, 1195, 3, 4], [467, 1199, 4, 3], [479, 1198, 4, 5], [486, 1196, 5, 8]], [[454, 138, 35, 49], [348, 142, 40, 50], [304, 166, 49, 67], [604, 142, 45, 58], [739, 153, 41, 68], [780, 164, 42, 57], [773, 210, 77, 85], [839, 323, 71, 154], [19, 56, 40, 43], [232, 205, 36, 74], [75, 269, 75, 88]], [[939, 455, 23, 28], [631, 359, 17, 25], [572, 401, 30, 35], [433, 395, 12, 18], [423, 392, 16, 21], [398, 391, 17, 21], [140, 393, 27, 29], [151, 337, 21, 29], [301, 389, 23, 21], [69, 312, 11, 12], [22, 307, 7, 8]], [[684, 165, 28, 38], [413, 397, 34, 48], [610, 355, 33, 45], [219, 175, 32, 40], [283, 187, 33, 43], [336, 189, 26, 39], [385, 193, 28, 38], [441, 178, 27, 38], [476, 216, 28, 45], [529, 155, 27, 39], [602, 166, 30, 38]], [[68, 209, 13, 15], [169, 200, 11, 14], [268, 219, 13, 15], [844, 203, 32, 45], [883, 30, 24, 39], [987, 73, 30, 45], [767, 10, 30, 38], [722, 25, 20, 32], [1000, 305, 23, 39], [830, 385, 35, 51], [703, 299, 32, 49]], [[1, 351, 117, 135], [103, 421, 53, 73], [420, 349, 21, 26], [553, 398, 84, 108], [730, 371, 91, 94], [699, 437, 28, 35], [612, 234, 29, 33], [858, 343, 16, 18], [932, 414, 14, 18], [912, 423, 21, 26], [977, 425, 43, 46]], [[722, 129, 57, 68], [588, 197, 50, 56], [530, 152, 42, 79], [382, 232, 38, 48], [293, 221, 33, 44], [230, 267, 30, 36], [195, 279, 23, 36], [161, 286, 23, 27], [0, 74, 154, 300], [99, 296, 19, 26], [27, 343, 83, 98]], [[738, 150, 42, 50], [581, 274, 48, 59], [519, 226, 51, 52], [475, 174, 53, 53], [424, 265, 48, 51], [387, 323, 44, 52], [270, 306, 35, 29], [328, 204, 43, 50], [283, 205, 24, 37], [211, 230, 42, 59], [706, 218, 44, 51]], [[616, 338, 40, 51], [803, 350, 32, 46], [77, 205, 28, 36], [38, 226, 19, 35], [166, 194, 34, 37], [255, 214, 34, 49], [301, 168, 34, 41], [357, 251, 27, 49], [389, 228, 29, 43], [464, 222, 34, 47], [578, 216, 27, 36]], [[64, 60, 42, 60], [224, 53, 43, 57], [398, 57, 44, 55], [589, 58, 42, 61], [753, 60, 39, 58], [945, 52, 36, 63], [904, 514, 36, 56], [737, 517, 45, 51], [579, 518, 42, 63], [417, 517, 38, 51], [231, 516, 40, 57], [69, 522, 41, 58]], [[864, 336, 19, 25], [790, 337, 19, 18], [566, 332, 23, 28], [501, 345, 20, 27], [467, 355, 19, 21], [403, 362, 11, 15], [238, 369, 9, 12], [205, 364, 15, 16], [182, 355, 17, 19], [128, 341, 21, 26], [106, 348, 16, 22], [62, 370, 13, 15]], [[527, 264, 39, 51], [618, 485, 22, 37], [792, 439, 41, 48], [984, 456, 33, 44], [1007, 472, 17, 36], [121, 726, 27, 34], [143, 746, 15, 22], [160, 729, 19, 24], [10, 407, 8, 12], [500, 2, 7, 8], [522, 2, 9, 8], [701, 23, 8, 10]], [[51, 192, 16, 19], [139, 186, 18, 23], [91, 216, 12, 21], [87, 328, 43, 56], [185, 370, 50, 80], [272, 191, 13, 20], [313, 187, 18, 19], [302, 277, 37, 45], [477, 255, 33, 43], [615, 238, 33, 43], [802, 219, 35, 41], [874, 310, 36, 74]], [[168, 459, 37, 49], [163, 385, 25, 63], [292, 605, 27, 49], [330, 575, 42, 58], [333, 623, 44, 59], [382, 554, 119, 167], [476, 423, 221, 319], [769, 469, 14, 21], [720, 597, 54, 65], [792, 567, 72, 96], [846, 600, 35, 59], [846, 464, 33, 134]], [[539, 69, 42, 51], [496, 121, 28, 37], [431, 151, 32, 33], [408, 142, 23, 34], [353, 131, 25, 28], [399, 121, 24, 27], [0, 118, 22, 38], [109, 37, 17, 24], [94, 44, 7, 9], [209, 107, 18, 22], [88, 153, 23, 25], [139, 125, 21, 34]], [[131, 508, 28, 66], [183, 475, 33, 58], [290, 507, 22, 32], [213, 683, 27, 54], [458, 440, 36, 45], [385, 282, 8, 13], [553, 284, 12, 13], [460, 312, 7, 9], [636, 226, 35, 45], [639, 657, 37, 44], [481, 353, 7, 10], [46, 365, 7, 18]], [[123, 111, 23, 41], [107, 49, 7, 12], [65, 52, 5, 10], [51, 58, 6, 7], [147, 45, 8, 12], [167, 41, 8, 15], [198, 41, 6, 9], [223, 39, 7, 13], [264, 36, 7, 13], [245, 40, 7, 10], [123, 52, 6, 9], [38, 61, 5, 7]], [[31, 410, 53, 68], [85, 421, 50, 65], [27, 597, 79, 76], [137, 412, 32, 63], [306, 612, 36, 72], [282, 245, 52, 76], [455, 281, 54, 72], [638, 479, 41, 75], [595, 161, 55, 88], [788, 173, 66, 88], [986, 574, 22, 38], [979, 309, 45, 34]], [[859, 175, 31, 32], [815, 164, 29, 33], [628, 230, 25, 29], [604, 375, 24, 24], [526, 290, 25, 33], [589, 232, 25, 31], [460, 206, 28, 35], [387, 317, 28, 38], [263, 260, 23, 28], [229, 199, 29, 30], [160, 121, 24, 29], [147, 173, 24, 26]], [[133, 460, 35, 53], [135, 407, 41, 52], [614, 616, 41, 67], [678, 530, 40, 64], [922, 507, 46, 42], [928, 205, 14, 14], [589, 188, 44, 58], [677, 232, 11, 13], [394, 230, 7, 9], [149, 228, 11, 10], [69, 233, 10, 9], [11, 224, 4, 6]], [[30, 112, 19, 20], [56, 104, 12, 16], [217, 100, 13, 16], [364, 123, 8, 11], [304, 118, 9, 11], [528, 134, 9, 7], [545, 135, 37, 45], [679, 139, 8, 11], [722, 149, 12, 11], [755, 155, 35, 31], [817, 69, 108, 135], [711, 143, 10, 14]], [[207, 277, 26, 30], [245, 256, 15, 19], [262, 240, 16, 19], [328, 260, 13, 16], [431, 269, 14, 16], [511, 260, 18, 20], [485, 263, 10, 14], [812, 267, 19, 24], [786, 251, 17, 20], [929, 275, 14, 15], [987, 274, 15, 17], [345, 410, 39, 54]], [[1003, 629, 16, 19], [948, 625, 18, 19], [850, 663, 58, 77], [742, 734, 41, 34], [618, 672, 56, 83], [550, 666, 28, 36], [447, 652, 29, 42], [399, 708, 52, 60], [41, 715, 45, 52], [147, 713, 31, 48], [100, 629, 48, 47], [126, 571, 16, 20]], [[775, 22, 71, 83], [958, 125, 36, 47], [532, 60, 34, 47], [492, 88, 27, 33], [293, 21, 63, 82], [375, 99, 21, 34], [200, 84, 18, 23], [168, 83, 19, 23], [57, 64, 19, 29], [119, 72, 17, 23], [8, 59, 19, 28], [902, 130, 31, 32]], [[281, 355, 35, 43], [161, 396, 37, 45], [1, 391, 34, 41], [732, 110, 48, 66], [905, 400, 25, 31], [853, 586, 33, 22], [726, 410, 33, 37], [612, 349, 37, 40], [660, 403, 27, 29], [451, 368, 33, 58], [453, 354, 36, 59], [365, 414, 46, 54]], [[103, 259, 49, 56], [308, 295, 50, 55], [445, 312, 44, 47], [573, 305, 46, 48], [602, 386, 38, 40], [759, 236, 51, 56], [798, 107, 44, 48], [657, 125, 43, 50], [502, 89, 38, 46], [359, 73, 46, 49], [222, 165, 39, 45], [97, 113, 47, 48]], [[211, 194, 45, 59], [304, 198, 40, 49], [364, 132, 56, 72], [422, 159, 28, 49], [116, 164, 46, 62], [66, 180, 42, 57], [615, 142, 47, 61], [512, 190, 37, 50], [764, 156, 47, 55], [978, 154, 36, 51], [820, 129, 54, 67], [529, 281, 110, 138]], [[109, 81, 18, 45], [403, 112, 97, 130], [512, 115, 108, 136], [695, 161, 30, 84], [715, 138, 35, 83], [717, 89, 35, 77], [846, 107, 8, 16], [902, 96, 16, 20], [908, 126, 10, 17], [948, 123, 15, 18], [1005, 122, 14, 21], [1010, 63, 7, 6]], [[5, 371, 74, 101], [483, 325, 93, 126], [637, 310, 99, 118], [711, 199, 31, 59], [542, 225, 42, 55], [14, 204, 76, 83], [329, 169, 28, 44], [49, 84, 30, 37], [849, 457, 48, 61], [94, 130, 33, 32], [580, 247, 21, 35], [463, 131, 47, 62]], [[542, 275, 50, 69], [788, 379, 18, 30], [873, 382, 11, 14], [922, 380, 12, 13], [326, 365, 13, 15], [298, 361, 12, 14], [200, 354, 11, 16], [157, 359, 11, 15], [109, 359, 13, 16], [63, 319, 13, 17], [37, 370, 14, 19], [22, 314, 14, 16]], [[178, 216, 35, 53], [387, 346, 18, 22], [599, 290, 9, 13], [635, 288, 12, 16], [636, 320, 10, 15], [605, 326, 9, 12], [882, 212, 30, 42], [702, 39, 95, 150], [891, 1, 114, 152], [414, 365, 12, 15], [622, 268, 9, 15], [437, 261, 8, 11]], [[860, 72, 48, 53], [794, 78, 34, 42], [789, 77, 25, 32], [750, 125, 24, 31], [452, 88, 41, 43], [485, 84, 25, 46], [533, 83, 27, 35], [635, 100, 32, 41], [113, 105, 38, 46], [172, 70, 37, 50], [215, 58, 31, 42], [287, 107, 34, 38]], [[919, 385, 88, 102], [723, 362, 68, 84], [497, 388, 60, 81], [381, 395, 70, 78], [193, 383, 77, 113], [28, 397, 94, 117], [220, 64, 66, 80], [355, 161, 54, 65], [473, 106, 59, 73], [646, 137, 51, 73], [808, 97, 60, 79], [21, 71, 11, 13]], [[835, 412, 37, 42], [767, 430, 39, 41], [877, 1014, 61, 65], [763, 969, 66, 84], [572, 399, 44, 63], [512, 430, 33, 41], [421, 411, 48, 52], [385, 422, 34, 55], [235, 530, 54, 78], [138, 434, 70, 78], [221, 943, 89, 115], [102, 906, 87, 103]], [[739, 381, 34, 44], [928, 622, 10, 12], [967, 545, 9, 13], [957, 554, 9, 11], [992, 557, 9, 13], [1014, 536, 10, 13], [853, 528, 15, 22], [942, 543, 8, 11], [871, 496, 6, 10], [142, 561, 11, 14], [100, 564, 9, 10], [115, 569, 9, 10]], [[208, 133, 103, 119], [630, 158, 23, 25], [758, 165, 15, 25], [899, 156, 20, 23], [924, 163, 14, 17], [962, 161, 16, 18], [666, 361, 28, 31], [653, 357, 17, 19], [602, 356, 10, 11], [577, 357, 9, 10], [673, 427, 19, 22], [413, 322, 25, 31]], [[140, 414, 14, 18], [154, 418, 9, 16], [186, 365, 10, 13], [236, 376, 7, 9], [211, 374, 8, 9], [84, 369, 8, 9], [245, 372, 7, 10], [70, 376, 5, 8], [75, 371, 8, 11], [37, 376, 6, 8], [23, 374, 8, 7], [271, 367, 6, 6]], [[623, 631, 16, 18], [123, 633, 12, 18], [21, 550, 6, 10], [261, 679, 12, 14], [423, 676, 7, 12], [467, 774, 7, 20], [578, 611, 8, 9], [465, 605, 5, 11], [409, 608, 4, 13], [653, 638, 6, 17], [739, 593, 6, 14], [987, 622, 7, 14], [705, 550, 6, 9]], [[323, 237, 46, 61], [390, 245, 39, 58], [413, 217, 38, 58], [423, 169, 42, 57], [459, 149, 31, 47], [665, 22, 32, 47], [618, 30, 31, 41], [594, 43, 36, 49], [539, 41, 43, 50], [552, 78, 38, 57], [678, 0, 30, 20], [189, 401, 30, 48], [539, 104, 18, 43]], [[309, 335, 20, 25], [418, 267, 45, 37], [590, 310, 15, 19], [606, 362, 10, 16], [644, 372, 18, 22], [675, 351, 10, 20], [715, 328, 10, 20], [832, 384, 10, 20], [886, 387, 9, 17], [931, 383, 11, 18], [973, 359, 14, 18], [1014, 385, 10, 24], [360, 147, 15, 17]], [[227, 492, 30, 35], [171, 592, 58, 72], [232, 563, 45, 57], [130, 514, 26, 36], [860, 505, 22, 31], [730, 560, 41, 43], [530, 547, 63, 90], [935, 486, 16, 22], [919, 541, 20, 25], [835, 508, 14, 12], [283, 497, 15, 23], [128, 585, 40, 59], [772, 623, 43, 60]], [[71, 237, 33, 62], [56, 254, 36, 52], [197, 290, 36, 50], [281, 279, 26, 36], [440, 244, 24, 35], [388, 240, 25, 38], [323, 286, 27, 36], [559, 262, 20, 34], [508, 553, 43, 57], [580, 472, 47, 52], [699, 267, 22, 30], [770, 263, 20, 27], [250, 258, 19, 22]], [[619, 289, 33, 65], [700, 360, 9, 12], [955, 434, 40, 47], [981, 413, 37, 55], [886, 452, 33, 42], [790, 473, 66, 114], [857, 420, 10, 12], [781, 434, 8, 10], [778, 417, 6, 7], [128, 445, 9, 10], [156, 433, 10, 13], [204, 424, 7, 8], [222, 423, 8, 8]], [[287, 234, 25, 41], [321, 238, 22, 37], [347, 241, 23, 41], [407, 243, 28, 40], [453, 236, 26, 39], [497, 244, 23, 40], [524, 248, 24, 41], [560, 227, 28, 40], [600, 249, 27, 40], [625, 242, 24, 40], [674, 246, 26, 39], [32, 369, 18, 38], [651, 235, 18, 33]], [[67, 94, 23, 31], [2, 76, 22, 30], [98, 54, 21, 26], [131, 51, 24, 29], [173, 92, 21, 30], [235, 51, 21, 32], [297, 77, 12, 21], [208, 109, 20, 20], [381, 96, 15, 20], [502, 88, 14, 19], [512, 111, 17, 18], [537, 218, 30, 40], [732, 71, 29, 29]], [[727, 177, 11, 16], [672, 159, 10, 18], [442, 131, 37, 45], [542, 116, 21, 31], [787, 153, 14, 18], [814, 164, 11, 15], [853, 173, 12, 16], [168, 156, 10, 14], [223, 130, 14, 19], [344, 152, 10, 16], [330, 125, 11, 22], [125, 143, 14, 16], [195, 147, 6, 9]], [[15, 135, 27, 48], [109, 129, 26, 47], [322, 192, 26, 51], [321, 106, 38, 50], [189, 86, 32, 48], [397, 105, 35, 46], [447, 170, 36, 45], [647, 127, 32, 48], [698, 197, 28, 57], [473, 125, 32, 45], [818, 177, 20, 40], [173, 149, 22, 42], [0, 164, 14, 50]], [[31, 28, 52, 68], [226, 21, 31, 28], [435, 15, 24, 27], [620, 36, 61, 77], [870, 15, 65, 84], [82, 288, 30, 33], [277, 295, 30, 33], [503, 285, 24, 26], [806, 276, 31, 41], [110, 556, 46, 62], [324, 560, 47, 57], [577, 511, 33, 41], [863, 525, 34, 38]], [[153, 106, 33, 42], [152, 151, 35, 50], [243, 127, 33, 36], [296, 149, 41, 45], [450, 164, 38, 35], [465, 119, 26, 30], [596, 119, 29, 34], [588, 171, 36, 41], [710, 131, 23, 25], [717, 191, 30, 38], [826, 131, 30, 27], [957, 144, 32, 38], [953, 191, 39, 36]], [[135, 389, 17, 18], [182, 383, 15, 23], [251, 396, 14, 14], [320, 403, 11, 17], [385, 386, 14, 19], [443, 403, 18, 20], [493, 384, 18, 17], [569, 405, 16, 19], [600, 404, 16, 17], [691, 396, 21, 21], [796, 419, 15, 20], [916, 423, 13, 17], [989, 366, 17, 19]], [[78, 612, 7, 14], [277, 571, 23, 34], [354, 484, 20, 26], [419, 518, 17, 30], [503, 509, 10, 25], [563, 566, 9, 23], [574, 562, 11, 27], [589, 560, 11, 27], [628, 571, 12, 26], [644, 574, 9, 22], [642, 459, 18, 66], [683, 452, 21, 79], [881, 352, 89, 143]], [[25, 353, 15, 19], [44, 373, 13, 13], [95, 374, 13, 20], [193, 375, 11, 19], [215, 375, 16, 19], [656, 450, 11, 12], [572, 429, 10, 11], [611, 447, 7, 11], [725, 454, 6, 10], [843, 474, 16, 22], [909, 478, 13, 14], [1009, 506, 9, 21], [752, 482, 11, 18]], [[706, 71, 24, 35], [754, 79, 29, 38], [841, 109, 29, 37], [193, 100, 32, 40], [237, 83, 26, 31], [293, 99, 31, 38], [331, 114, 28, 35], [392, 148, 26, 26], [416, 92, 31, 36], [498, 75, 29, 41], [566, 73, 28, 31], [587, 104, 27, 39], [651, 75, 27, 36]], [[903, 278, 33, 45], [735, 248, 52, 64], [576, 236, 54, 64], [473, 284, 41, 42], [376, 221, 48, 51], [459, 251, 33, 35], [345, 264, 29, 37], [207, 238, 44, 42], [120, 185, 53, 55], [12, 162, 52, 64], [434, 24, 70, 92], [184, 0, 91, 116], [0, 0, 61, 77]], [[752, 247, 16, 18], [646, 285, 21, 17], [790, 303, 16, 14], [675, 326, 21, 15], [885, 381, 10, 10], [889, 393, 9, 12], [38, 327, 6, 11], [46, 324, 8, 10], [22, 319, 9, 10], [137, 357, 5, 9], [121, 337, 5, 7], [153, 331, 5, 8], [101, 308, 6, 8]], [[114, 314, 20, 22], [184, 315, 9, 12], [129, 293, 13, 15], [237, 312, 10, 14], [866, 368, 13, 11], [919, 354, 9, 10], [213, 315, 7, 12], [271, 316, 10, 12], [6, 299, 9, 20], [155, 309, 10, 16], [977, 352, 4, 6], [317, 320, 8, 12], [62, 314, 12, 16]], [[708, 251, 100, 142], [332, 144, 63, 83], [411, 203, 29, 46], [608, 239, 56, 74], [393, 379, 89, 121], [969, 169, 55, 115], [668, 215, 63, 97], [153, 220, 33, 53], [47, 225, 25, 50], [85, 250, 26, 24], [475, 190, 49, 62], [554, 254, 19, 25], [946, 234, 9, 11]], [[31, 121, 17, 19], [83, 175, 31, 39], [120, 198, 40, 46], [199, 172, 49, 57], [307, 98, 31, 33], [307, 124, 39, 55], [435, 61, 38, 44], [492, 146, 28, 40], [597, 151, 85, 104], [866, 33, 41, 53], [180, 152, 5, 7], [163, 164, 4, 7], [144, 154, 5, 7]], [[0, 141, 13, 67], [108, 245, 50, 63], [234, 216, 33, 43], [286, 208, 53, 70], [351, 184, 40, 60], [489, 145, 59, 79], [608, 130, 45, 62], [656, 112, 35, 63], [677, 115, 56, 95], [765, 145, 45, 57], [859, 169, 38, 39], [857, 217, 55, 60], [34, 357, 44, 89]], [[82, 262, 23, 28], [147, 287, 35, 43], [280, 265, 21, 27], [347, 255, 29, 37], [448, 275, 23, 34], [495, 281, 30, 39], [555, 254, 24, 25], [565, 298, 40, 60], [719, 236, 23, 27], [746, 237, 32, 37], [967, 255, 31, 41], [639, 254, 29, 40], [573, 419, 52, 35]], [[535, 278, 56, 61], [817, 107, 34, 44], [878, 19, 16, 23], [839, 0, 16, 21], [763, 45, 22, 27], [473, 34, 24, 26], [526, 12, 19, 22], [360, 72, 28, 33], [366, 3, 16, 21], [323, 18, 18, 24], [148, 48, 25, 28], [23, 22, 17, 23], [162, 4, 16, 20]], [[681, 234, 11, 11], [744, 380, 14, 18], [850, 416, 16, 16], [907, 354, 16, 22], [918, 376, 24, 32], [968, 405, 22, 27], [1005, 416, 14, 20], [800, 425, 14, 16], [621, 391, 16, 22], [452, 349, 22, 23], [529, 311, 18, 25], [227, 428, 23, 36], [717, 416, 11, 14]], [[219, 87, 49, 61], [355, 148, 36, 51], [345, 228, 44, 65], [373, 70, 36, 49], [418, 113, 40, 56], [482, 61, 36, 52], [536, 158, 41, 56], [567, 63, 39, 50], [659, 89, 39, 50], [640, 282, 47, 51], [789, 149, 41, 51], [829, 86, 41, 54], [1001, 71, 8, 39]], [[567, 199, 40, 47], [711, 225, 37, 41], [823, 204, 29, 39], [904, 207, 29, 37], [166, 348, 20, 24], [826, 98, 10, 11], [803, 69, 8, 8], [761, 73, 10, 13], [721, 83, 11, 11], [708, 48, 11, 15], [823, 45, 10, 11], [852, 90, 11, 13], [450, 209, 39, 49]], [[433, 461, 12, 15], [351, 456, 10, 9], [415, 446, 7, 9], [436, 435, 6, 9], [462, 449, 8, 10], [503, 443, 7, 8], [520, 454, 10, 11], [541, 457, 10, 13], [555, 449, 12, 15], [586, 440, 13, 15], [615, 445, 12, 12], [668, 445, 11, 13], [702, 437, 10, 11]], [[56, 395, 12, 20], [108, 401, 17, 26], [82, 427, 24, 37], [170, 390, 26, 38], [218, 405, 19, 26], [382, 420, 12, 24], [409, 422, 12, 23], [431, 415, 14, 24], [477, 414, 20, 27], [538, 369, 34, 46], [584, 413, 17, 25], [598, 415, 8, 21], [910, 421, 11, 17], [886, 385, 19, 27]], [[836, 20, 18, 21], [793, 81, 17, 18], [728, 67, 16, 32], [739, 96, 14, 18], [679, 132, 12, 18], [322, 340, 15, 14], [314, 365, 12, 15], [234, 393, 16, 14], [224, 413, 13, 13], [123, 482, 12, 11], [79, 472, 15, 17], [575, 137, 16, 33], [568, 172, 13, 22], [153, 384, 21, 33]], [[547, 102, 44, 76], [601, 111, 48, 62], [21, 402, 21, 27], [96, 403, 22, 26], [150, 398, 19, 24], [196, 391, 19, 24], [256, 385, 15, 22], [39, 383, 17, 22], [131, 378, 14, 21], [57, 364, 16, 20], [103, 353, 17, 21], [176, 371, 17, 20], [228, 370, 16, 20], [307, 320, 19, 24]], [[977, 196, 23, 27], [680, 246, 51, 50], [530, 202, 41, 42], [635, 210, 41, 47], [605, 224, 17, 20], [765, 241, 35, 51], [365, 230, 42, 54], [447, 201, 35, 39], [344, 169, 31, 36], [260, 151, 31, 38], [226, 198, 36, 54], [108, 143, 31, 37], [156, 156, 17, 17], [4, 160, 30, 32]], [[409, 208, 23, 31], [276, 170, 24, 30], [588, 129, 32, 44], [777, 336, 40, 39], [949, 271, 29, 40], [213, 280, 38, 51], [970, 104, 6, 8], [943, 105, 6, 8], [962, 105, 5, 8], [922, 97, 6, 8], [937, 97, 6, 8], [891, 96, 5, 8], [879, 96, 5, 8], [997, 106, 8, 10]], [[267, 186, 54, 59], [45, 105, 61, 71], [186, 100, 65, 85], [336, 159, 59, 70], [398, 158, 42, 57], [449, 190, 56, 62], [599, 143, 60, 69], [730, 120, 61, 69], [859, 176, 72, 80], [523, 227, 39, 47], [661, 187, 45, 52], [0, 181, 35, 55], [922, 217, 40, 56], [493, 239, 33, 48]], [[103, 315, 77, 82], [124, 166, 52, 76], [169, 206, 34, 48], [377, 157, 32, 38], [466, 162, 37, 41], [429, 40, 31, 38], [623, 190, 54, 68], [706, 199, 67, 86], [720, 295, 115, 157], [820, 123, 14, 13], [866, 101, 15, 25], [746, 98, 8, 20], [199, 117, 20, 19], [377, 109, 20, 22]], [[140, 0, 33, 20], [233, 0, 22, 22], [310, 0, 28, 20], [103, 197, 38, 44], [222, 160, 39, 50], [429, 170, 26, 36], [353, 208, 36, 41], [399, 221, 38, 42], [506, 184, 43, 42], [590, 190, 27, 29], [662, 194, 38, 42], [768, 157, 41, 38], [843, 187, 32, 40], [921, 211, 35, 41]], [[78, 159, 32, 40], [183, 156, 32, 40], [131, 314, 34, 49], [222, 309, 35, 51], [289, 158, 34, 49], [398, 151, 34, 44], [383, 303, 35, 50], [507, 144, 31, 43], [506, 330, 32, 47], [603, 144, 32, 38], [637, 297, 35, 42], [715, 123, 31, 41], [814, 159, 33, 44], [779, 273, 38, 45]], [[860, 303, 35, 38], [736, 321, 36, 40], [597, 308, 39, 39], [985, 338, 11, 16], [981, 295, 10, 13], [393, 312, 30, 34], [472, 304, 32, 37], [295, 319, 30, 35], [308, 291, 26, 35], [518, 286, 33, 37], [26, 333, 13, 13], [176, 319, 32, 40], [87, 315, 29, 34], [409, 288, 33, 38]], [[872, 33, 55, 73], [734, 48, 57, 70], [599, 28, 59, 68], [460, 76, 56, 69], [369, 54, 58, 74], [268, 49, 56, 68], [99, 38, 53, 71], [90, 276, 59, 69], [209, 314, 46, 69], [380, 301, 53, 63], [536, 299, 58, 69], [653, 276, 54, 64], [813, 283, 51, 66], [921, 278, 61, 76]], [[547, 174, 18, 20], [437, 163, 17, 23], [421, 164, 14, 23], [389, 163, 14, 23], [344, 140, 18, 24], [276, 141, 13, 20], [350, 172, 11, 17], [866, 223, 20, 27], [859, 213, 12, 17], [927, 213, 12, 16], [189, 124, 31, 25], [477, 167, 21, 25], [208, 142, 15, 20], [7, 131, 22, 26]], [[290, 119, 40, 59], [66, 488, 16, 23], [88, 507, 9, 11], [102, 519, 11, 14], [138, 528, 9, 11], [110, 515, 5, 12], [157, 527, 8, 11], [178, 527, 10, 13], [194, 532, 8, 12], [225, 522, 14, 17], [370, 629, 31, 45], [422, 568, 10, 16], [455, 570, 9, 12], [497, 578, 14, 16]], [[1002, 474, 19, 22], [921, 479, 15, 20], [884, 482, 20, 25], [859, 489, 16, 16], [718, 490, 17, 21], [744, 424, 15, 22], [602, 465, 17, 28], [578, 400, 20, 25], [471, 425, 16, 23], [404, 507, 40, 54], [380, 420, 22, 27], [493, 480, 14, 18], [63, 339, 22, 36], [244, 418, 20, 28]], [[384, 331, 32, 39], [948, 189, 68, 104], [514, 296, 36, 34], [562, 272, 27, 39], [587, 235, 24, 32], [697, 253, 17, 24], [742, 267, 14, 19], [815, 260, 14, 18], [891, 257, 10, 13], [890, 310, 11, 16], [638, 238, 22, 28], [242, 254, 11, 11], [859, 274, 8, 11], [868, 269, 10, 14]], [[955, 77, 26, 31], [866, 72, 30, 31], [834, 37, 21, 26], [773, 97, 27, 30], [764, 38, 12, 17], [740, 36, 12, 17], [669, 29, 16, 24], [556, 26, 20, 26], [467, 21, 20, 25], [427, 68, 19, 21], [282, 46, 39, 49], [193, 98, 43, 54], [159, 14, 38, 47], [58, 48, 44, 52]], [[63, 81, 182, 140], [743, 251, 55, 57], [905, 247, 57, 71], [252, 644, 23, 27], [256, 605, 23, 27], [111, 728, 65, 69], [421, 752, 33, 46], [468, 751, 37, 52], [509, 779, 38, 61], [550, 811, 34, 47], [423, 900, 29, 63], [625, 872, 41, 57], [737, 896, 47, 63], [84, 670, 15, 29]], [[73, 175, 45, 47], [200, 185, 37, 48], [240, 367, 45, 45], [328, 312, 36, 42], [316, 234, 33, 41], [386, 295, 34, 41], [426, 394, 43, 47], [526, 328, 32, 36], [594, 382, 41, 46], [653, 336, 38, 42], [770, 365, 43, 49], [800, 197, 32, 40], [847, 198, 31, 37], [949, 383, 45, 45]], [[870, 339, 23, 23], [422, 378, 21, 24], [148, 219, 34, 34], [220, 223, 28, 35], [345, 236, 25, 34], [394, 250, 25, 31], [437, 240, 26, 29], [452, 280, 25, 29], [515, 253, 20, 25], [631, 200, 37, 40], [701, 229, 34, 36], [816, 229, 24, 29], [859, 204, 29, 33], [805, 291, 25, 27]], [[208, 416, 10, 16], [242, 386, 12, 14], [102, 429, 16, 19], [239, 413, 15, 13], [284, 418, 10, 12], [367, 414, 8, 12], [407, 409, 7, 13], [500, 405, 6, 9], [450, 414, 5, 9], [784, 319, 5, 5], [754, 317, 5, 6], [634, 319, 5, 6], [968, 331, 7, 7], [953, 332, 6, 6]], [[200, 68, 42, 58], [298, 105, 40, 55], [462, 90, 34, 50], [203, 262, 48, 63], [396, 257, 38, 50], [518, 249, 40, 52], [627, 405, 44, 56], [489, 427, 43, 72], [353, 410, 44, 60], [715, 88, 34, 49], [580, 40, 38, 46], [754, 233, 38, 58], [640, 247, 38, 51], [772, 409, 44, 57]], [[277, 89, 76, 108], [957, 783, 53, 67], [99, 781, 66, 76], [70, 850, 72, 62], [80, 910, 75, 84], [246, 790, 50, 63], [276, 859, 71, 95], [357, 832, 61, 67], [369, 772, 48, 60], [520, 808, 53, 76], [645, 930, 55, 66], [685, 775, 53, 65], [809, 796, 65, 86], [860, 783, 46, 67]], [[409, 212, 27, 39], [492, 156, 14, 22], [414, 151, 12, 18], [442, 151, 9, 21], [184, 200, 25, 35], [118, 167, 23, 26], [218, 212, 21, 36], [8, 149, 17, 29], [374, 154, 9, 15], [222, 136, 9, 11], [271, 178, 7, 11], [622, 213, 33, 34], [879, 203, 31, 35], [453, 154, 10, 14]], [[98, 349, 37, 46], [216, 336, 37, 43], [342, 335, 41, 49], [484, 331, 38, 52], [664, 361, 42, 48], [857, 345, 43, 52], [814, 138, 33, 40], [710, 148, 40, 45], [615, 137, 35, 49], [521, 150, 38, 45], [408, 169, 37, 42], [315, 145, 35, 45], [218, 139, 31, 39], [143, 215, 3, 4]], [[160, 456, 54, 64], [114, 205, 25, 28], [413, 321, 29, 39], [429, 239, 19, 29], [494, 279, 41, 48], [520, 237, 20, 25], [529, 159, 21, 30], [599, 274, 31, 45], [639, 264, 48, 52], [776, 246, 56, 56], [665, 155, 21, 24], [763, 182, 15, 27], [374, 297, 26, 17], [926, 303, 38, 43]], [[209, 73, 16, 26], [461, 86, 18, 22], [379, 85, 15, 23], [312, 53, 17, 27], [97, 56, 16, 25], [60, 79, 15, 21], [176, 81, 20, 24], [554, 76, 15, 23], [641, 64, 14, 20], [702, 63, 15, 26], [801, 89, 14, 24], [893, 82, 13, 21], [917, 81, 14, 23], [981, 71, 14, 20]], [[16, 128, 22, 27], [68, 133, 19, 26], [91, 132, 21, 25], [113, 140, 13, 16], [125, 151, 19, 21], [163, 143, 20, 21], [203, 124, 21, 28], [277, 113, 25, 32], [381, 111, 25, 29], [404, 92, 29, 37], [514, 79, 35, 41], [614, 94, 29, 37], [718, 98, 34, 45], [912, 113, 36, 43]], [[799, 245, 65, 76], [586, 201, 43, 47], [814, 129, 9, 12], [831, 114, 10, 12], [903, 129, 12, 14], [866, 118, 6, 8], [895, 111, 7, 9], [928, 110, 8, 9], [997, 117, 9, 12], [958, 109, 6, 9], [799, 129, 9, 10], [793, 111, 6, 8], [425, 137, 11, 12], [511, 121, 7, 9]], [[848, 236, 19, 23], [934, 208, 7, 8], [724, 231, 15, 22], [644, 227, 15, 23], [649, 223, 23, 28], [606, 245, 16, 23], [579, 234, 26, 32], [611, 224, 6, 6], [488, 247, 29, 31], [411, 255, 27, 32], [354, 289, 30, 31], [96, 181, 48, 67], [187, 259, 33, 26], [283, 246, 29, 31], [0, 319, 17, 32]], [[530, 1181, 94, 123], [313, 1300, 63, 70], [543, 2046, 234, 190], [259, 2478, 80, 83], [79, 2466, 78, 95], [640, 3345, 128, 146], [419, 3543, 127, 92], [920, 4315, 104, 126], [732, 478, 121, 151], [384, 537, 28, 44], [354, 489, 45, 61], [275, 500, 51, 48], [206, 491, 43, 45], [164, 522, 43, 56], [73, 499, 40, 58]], [[1, 269, 36, 38], [37, 251, 31, 38], [122, 251, 30, 40], [178, 248, 30, 33], [291, 223, 32, 38], [432, 253, 27, 37], [508, 268, 23, 28], [558, 242, 26, 31], [613, 256, 21, 31], [680, 256, 19, 26], [718, 273, 13, 20], [942, 248, 31, 31], [896, 246, 30, 29], [826, 255, 29, 26], [761, 257, 27, 30]], [[959, 253, 20, 24], [907, 242, 24, 26], [844, 215, 24, 30], [784, 233, 22, 25], [718, 256, 23, 28], [655, 253, 23, 29], [583, 243, 25, 26], [523, 256, 24, 26], [431, 255, 22, 24], [365, 239, 23, 27], [296, 237, 23, 26], [235, 239, 23, 23], [164, 221, 24, 28], [101, 228, 24, 25], [40, 249, 20, 25]], [[835, 94, 46, 63], [883, 0, 53, 28], [579, 169, 41, 48], [785, 341, 60, 68], [675, 338, 54, 54], [445, 343, 50, 67], [274, 348, 55, 72], [594, 376, 47, 58], [148, 328, 46, 57], [894, 686, 69, 109], [531, 652, 63, 85], [766, 545, 35, 50], [401, 360, 42, 65], [35, 336, 50, 63], [772, 118, 36, 48]], [[785, 211, 18, 27], [506, 215, 18, 25], [470, 232, 18, 27], [421, 268, 14, 21], [297, 259, 18, 25], [666, 259, 11, 22], [517, 262, 13, 19], [527, 257, 13, 21], [6, 294, 27, 33], [595, 254, 15, 26], [553, 271, 11, 19], [652, 251, 11, 22], [691, 258, 12, 20], [864, 249, 14, 20], [431, 247, 14, 15]], [[15, 41, 26, 24], [89, 65, 22, 21], [135, 41, 22, 30], [201, 42, 22, 21], [249, 54, 19, 22], [306, 43, 25, 26], [398, 31, 23, 29], [488, 34, 20, 24], [602, 41, 21, 26], [685, 58, 22, 22], [763, 47, 21, 22], [824, 54, 18, 23], [937, 44, 26, 28], [910, 55, 22, 23], [1013, 58, 11, 27]], [[704, 42, 18, 38], [583, 91, 26, 30], [869, 152, 22, 34], [887, 173, 25, 48], [717, 188, 37, 36], [529, 156, 29, 37], [518, 219, 31, 37], [665, 257, 27, 61], [932, 420, 92, 146], [769, 332, 40, 76], [366, 84, 20, 32], [290, 96, 16, 24], [198, 89, 15, 24], [34, 80, 22, 21], [168, 232, 31, 58]], [[47, 90, 24, 24], [127, 119, 26, 24], [274, 99, 22, 21], [300, 97, 27, 25], [388, 96, 28, 29], [442, 95, 27, 28], [500, 74, 32, 33], [617, 109, 18, 21], [731, 93, 24, 26], [849, 99, 24, 27], [822, 117, 23, 25], [894, 102, 16, 18], [928, 112, 20, 21], [959, 91, 18, 18], [195, 107, 20, 22]], [[190, 177, 74, 117], [144, 36, 21, 26], [631, 60, 17, 25], [668, 46, 16, 21], [696, 63, 21, 21], [730, 59, 15, 19], [607, 53, 17, 15], [586, 52, 14, 21], [525, 50, 18, 24], [475, 50, 14, 18], [428, 39, 15, 21], [826, 54, 16, 22], [856, 56, 15, 17], [894, 53, 12, 12], [986, 53, 11, 14]], [[586, 211, 16, 20], [442, 206, 14, 20], [48, 286, 14, 19], [20, 310, 11, 15], [927, 361, 13, 21], [958, 431, 17, 29], [195, 375, 14, 24], [121, 370, 14, 23], [82, 433, 12, 27], [166, 428, 13, 26], [82, 481, 19, 33], [799, 444, 13, 26], [812, 374, 14, 28], [920, 231, 12, 17], [957, 221, 13, 19]], [[140, 318, 24, 29], [161, 282, 21, 22], [149, 404, 89, 105], [236, 278, 29, 34], [261, 275, 35, 36], [308, 394, 34, 40], [353, 293, 41, 58], [517, 300, 36, 34], [490, 289, 24, 23], [461, 273, 10, 11], [662, 281, 37, 45], [744, 226, 57, 72], [986, 269, 38, 46], [882, 241, 30, 43], [59, 364, 39, 46]], [[293, 322, 64, 63], [345, 244, 57, 57], [463, 157, 33, 35], [383, 163, 34, 39], [439, 158, 22, 32], [353, 166, 21, 31], [543, 145, 25, 32], [721, 137, 28, 35], [782, 128, 33, 38], [820, 132, 24, 32], [608, 199, 25, 33], [281, 28, 43, 49], [655, 147, 20, 33], [628, 152, 24, 36], [561, 161, 26, 45]], [[130, 160, 70, 99], [242, 327, 23, 25], [356, 304, 26, 35], [581, 266, 16, 23], [526, 275, 13, 16], [485, 287, 18, 19], [507, 299, 19, 17], [619, 264, 17, 22], [668, 255, 12, 17], [725, 243, 12, 14], [877, 281, 12, 12], [837, 269, 15, 21], [849, 298, 21, 25], [821, 310, 19, 27], [639, 327, 19, 27]], [[267, 205, 26, 34], [124, 211, 27, 36], [366, 236, 27, 35], [422, 220, 24, 34], [492, 249, 24, 28], [452, 245, 20, 24], [390, 239, 20, 21], [554, 250, 22, 30], [726, 249, 36, 34], [646, 269, 23, 24], [866, 280, 24, 24], [891, 280, 17, 19], [935, 261, 26, 43], [347, 228, 21, 27], [193, 199, 35, 37]], [[792, 1027, 35, 49], [242, 1176, 22, 33], [319, 1217, 11, 14], [102, 1204, 13, 19], [10, 1201, 12, 17], [79, 1211, 8, 11], [405, 1208, 8, 11], [186, 1211, 8, 12], [777, 648, 53, 67], [657, 639, 62, 75], [271, 599, 60, 81], [636, 199, 12, 16], [883, 150, 8, 11], [392, 200, 10, 15], [341, 220, 14, 27]], [[285, 91, 4, 6], [311, 95, 3, 6], [341, 66, 4, 6], [344, 93, 5, 5], [359, 96, 5, 6], [328, 97, 5, 8], [236, 124, 3, 3], [232, 118, 4, 4], [649, 320, 27, 34], [598, 285, 14, 26], [828, 276, 13, 19], [951, 348, 26, 41], [276, 175, 8, 9], [475, 181, 8, 9], [64, 212, 9, 11]], [[220, 161, 34, 42], [150, 173, 31, 36], [94, 175, 22, 29], [120, 190, 24, 35], [10, 230, 17, 23], [300, 344, 21, 21], [579, 188, 31, 35], [467, 188, 30, 34], [392, 212, 22, 28], [322, 232, 18, 26], [404, 218, 24, 28], [669, 214, 24, 29], [743, 215, 28, 32], [957, 179, 30, 39], [830, 212, 26, 33]], [[804, 127, 37, 37], [931, 75, 19, 25], [955, 67, 16, 30], [977, 94, 19, 30], [1000, 96, 11, 30], [996, 22, 11, 17], [912, 27, 14, 21], [906, 66, 13, 23], [767, 0, 26, 18], [623, 85, 31, 38], [930, 0, 17, 19], [150, 171, 30, 44], [340, 104, 30, 36], [204, 68, 30, 39], [449, 40, 28, 38]], [[125, 366, 15, 24], [189, 353, 15, 24], [210, 369, 13, 22], [220, 149, 12, 16], [295, 143, 15, 20], [296, 360, 16, 22], [377, 150, 13, 17], [463, 141, 13, 18], [475, 350, 17, 29], [578, 340, 13, 28], [627, 324, 17, 27], [660, 336, 14, 22], [724, 344, 15, 20], [815, 316, 22, 32], [958, 334, 15, 20]], [[56, 175, 43, 57], [104, 150, 38, 38], [196, 118, 29, 37], [214, 139, 43, 48], [295, 132, 41, 51], [340, 138, 48, 63], [391, 132, 38, 58], [470, 163, 48, 74], [491, 104, 50, 57], [535, 87, 36, 62], [601, 199, 49, 61], [785, 119, 47, 76], [835, 119, 55, 82], [867, 88, 47, 62], [1009, 115, 15, 59]], [[65, 300, 34, 41], [134, 332, 14, 15], [199, 343, 7, 10], [262, 327, 9, 12], [320, 336, 16, 20], [356, 300, 30, 35], [480, 319, 20, 26], [540, 290, 33, 39], [571, 292, 25, 34], [630, 313, 25, 30], [387, 310, 22, 28], [750, 276, 33, 46], [815, 323, 19, 23], [936, 326, 20, 21], [1014, 310, 9, 19]], [[60, 412, 32, 67], [91, 376, 30, 49], [217, 350, 23, 27], [195, 420, 59, 84], [357, 347, 24, 39], [303, 348, 23, 40], [309, 325, 20, 26], [422, 312, 11, 24], [478, 334, 18, 28], [464, 309, 13, 21], [708, 175, 18, 22], [703, 286, 12, 22], [540, 326, 14, 32], [344, 590, 68, 168], [574, 381, 12, 50]], [[133, 403, 27, 40], [233, 377, 25, 30], [335, 366, 25, 29], [389, 367, 23, 29], [468, 350, 15, 31], [544, 354, 20, 21], [610, 345, 21, 26], [639, 342, 20, 21], [718, 335, 18, 25], [808, 347, 18, 23], [849, 376, 22, 29], [764, 400, 18, 29], [927, 378, 17, 25], [666, 400, 17, 27], [561, 430, 15, 26]], [[67, 149, 36, 46], [73, 211, 36, 68], [166, 229, 57, 74], [353, 231, 64, 82], [388, 132, 19, 52], [579, 144, 37, 64], [606, 120, 27, 52], [792, 134, 25, 42], [806, 136, 25, 40], [944, 121, 18, 33], [870, 146, 19, 27], [925, 199, 48, 67], [712, 338, 48, 50], [154, 390, 69, 91], [238, 149, 12, 31]], [[194, 226, 24, 28], [269, 268, 20, 23], [334, 280, 14, 18], [385, 297, 11, 15], [154, 298, 15, 16], [37, 266, 18, 23], [603, 237, 28, 25], [929, 248, 23, 20], [840, 288, 17, 16], [780, 298, 12, 10], [1013, 268, 11, 22], [822, 269, 10, 13], [386, 281, 7, 8], [437, 312, 8, 11], [409, 309, 10, 12]], [[846, 146, 91, 98], [976, 111, 48, 131], [783, 124, 72, 72], [742, 142, 53, 69], [648, 77, 92, 108], [943, 153, 53, 67], [508, 142, 120, 147], [388, 111, 106, 132], [359, 160, 71, 76], [253, 166, 80, 88], [317, 152, 41, 40], [101, 74, 139, 151], [19, 166, 70, 79], [0, 179, 51, 90], [738, 101, 24, 27]], [[166, 374, 5, 7], [461, 357, 9, 19], [698, 436, 10, 13], [675, 399, 7, 11], [704, 399, 7, 10], [620, 394, 9, 13], [599, 397, 7, 9], [633, 400, 6, 9], [572, 461, 6, 11], [823, 395, 8, 12], [873, 403, 6, 8], [905, 408, 7, 10], [963, 409, 7, 10], [957, 396, 9, 11], [920, 411, 10, 10]]]
