--- models/deconv.py
@@ -1,45 +1,0 @@
-num_deconv_layers=3,
-num_deconv_filters=(256, 256, 256),
-num_deconv_kernels=(4, 4, 4),
-
-if num_deconv_layers > 0:
-    self.deconv_layers = self._make_deconv_layer(
-    num_deconv_layers,
-    num_deconv_filters,
-    num_deconv_kernels,
-)
-
-def _make_deconv_layer(self, num_layers, num_filters, num_kernels):
-        """Make deconv layers."""
-        if num_layers != len(num_filters):
-            error_msg = f'num_layers({num_layers}) ' \
-                        f'!= length of num_filters({len(num_filters)})'
-            raise ValueError(error_msg)
-        if num_layers != len(num_kernels):
-            error_msg = f'num_layers({num_layers}) ' \
-                        f'!= length of num_kernels({len(num_kernels)})'
-            raise ValueError(error_msg)
-
-        layers = []
-        for i in range(num_layers):
-            kernel, padding, output_padding = \
-                self._get_deconv_cfg(num_kernels[i])
-
-            planes = num_filters[i]
-            layers.append(
-                build_upsample_layer(
-                    dict(type='deconv'),
-                    in_channels=self.in_channels,
-                    out_channels=planes,
-                    kernel_size=kernel,
-                    stride=2,
-                    padding=padding,
-                    output_padding=output_padding,
-                    bias=False))
-            layers.append(nn.BatchNorm2d(planes))
-            layers.append(nn.ReLU(inplace=True))
-            self.in_channels = planes
-
-        return nn.Sequential(*layers)
-
-
