--- test_data.py
@@ -1,2 +1,0 @@
-test_data = ['12--Group/12_Group_Large_Group_12_Group_Large_Group_12_551.jpg', '41--Swimming/41_Swimming_Swimmer_41_376.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_638.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_888.jpg', '38--Tennis/38_Tennis_Tennis_38_604.jpg', '46--Jockey/46_Jockey_Jockey_46_728.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_231.jpg', '52--Photographers/52_Photographers_photographertakingphoto_52_456.jpg', '41--Swimming/41_Swimming_Swimmer_41_170.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_144.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_353.jpg', '57--Angler/57_Angler_peoplefishing_57_515.jpg', '41--Swimming/41_Swimming_Swimmer_41_262.jpg', '57--Angler/57_Angler_peoplefishing_57_250.jpg', '46--Jockey/46_Jockey_Jockey_46_393.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_735.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_767.jpg', '34--Baseball/34_Baseball_Baseball_34_580.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_771.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_197.jpg', '35--Basketball/35_Basketball_Basketball_35_791.jpg', '59--people--driving--car/59_peopledrivingcar_peopledrivingcar_59_725.jpg', '35--Basketball/35_Basketball_playingbasketball_35_730.jpg', '18--Concerts/18_Concerts_Concerts_18_1004.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_860.jpg', '59--people--driving--car/59_peopledrivingcar_peopledrivingcar_59_978.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_728.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_312.jpg', '7--Cheering/7_Cheering_Cheering_7_239.jpg', '19--Couple/19_Couple_Couple_19_835.jpg', '57--Angler/57_Angler_peoplefishing_57_402.jpg', '3--Riot/3_Riot_Riot_3_506.jpg', '27--Spa/27_Spa_Spa_27_121.jpg', '43--Row_Boat/43_Row_Boat_Canoe_43_726.jpg', '7--Cheering/7_Cheering_Cheering_7_558.jpg', '46--Jockey/46_Jockey_Jockey_46_188.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_996.jpg', '5--Car_Accident/5_Car_Accident_Car_Crash_5_457.jpg', '1--Handshaking/1_Handshaking_Handshaking_1_762.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_666.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_595.jpg', '43--Row_Boat/43_Row_Boat_Rowboat_43_341.jpg', '5--Car_Accident/5_Car_Accident_Car_Crash_5_451.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_327.jpg', '39--Ice_Skating/39_Ice_Skating_iceskiing_39_121.jpg', '16--Award_Ceremony/16_Award_Ceremony_Awards_Ceremony_16_270.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_Matador_Bullfighter_47_385.jpg', '35--Basketball/35_Basketball_playingbasketball_35_135.jpg', '28--Sports_Fan/28_Sports_Fan_Sports_Fan_28_265.jpg', '4--Dancing/4_Dancing_Dancing_4_983.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_345.jpg', '51--Dresses/51_Dresses_wearingdress_51_869.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_499.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_153.jpg', '16--Award_Ceremony/16_Award_Ceremony_Awards_Ceremony_16_56.jpg', '4--Dancing/4_Dancing_Dancing_4_1036.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_170.jpg', '51--Dresses/51_Dresses_wearingdress_51_150.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_304.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_987.jpg', '37--Soccer/37_Soccer_soccer_ball_37_233.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_227.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_491.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_438.jpg', '28--Sports_Fan/28_Sports_Fan_Sports_Fan_28_868.jpg', '19--Couple/19_Couple_Couple_19_125.jpg', '35--Basketball/35_Basketball_playingbasketball_35_582.jpg', '33--Running/33_Running_Running_33_119.jpg', '37--Soccer/37_Soccer_soccer_ball_37_926.jpg', '36--Football/36_Football_americanfootball_ball_36_38.jpg', '46--Jockey/46_Jockey_Jockey_46_54.jpg', '30--Surgeons/30_Surgeons_Surgeons_30_552.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_212.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_759.jpg', '46--Jockey/46_Jockey_Jockey_46_933.jpg', '6--Funeral/6_Funeral_Funeral_6_1006.jpg', '34--Baseball/34_Baseball_Baseball_34_828.jpg', '37--Soccer/37_Soccer_soccer_ball_37_818.jpg', '53--Raid/53_Raid_policeraid_53_770.jpg', '4--Dancing/4_Dancing_Dancing_4_1029.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_281.jpg', '54--Rescue/54_Rescue_rescuepeople_54_711.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_463.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_412.jpg', '35--Basketball/35_Basketball_Basketball_35_458.jpg', '30--Surgeons/30_Surgeons_Surgeons_30_823.jpg', '51--Dresses/51_Dresses_wearingdress_51_672.jpg', '43--Row_Boat/43_Row_Boat_Rowboat_43_907.jpg', '51--Dresses/51_Dresses_wearingdress_51_549.jpg', '22--Picnic/22_Picnic_Picnic_22_10.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_658.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_430.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_318.jpg', '13--Interview/13_Interview_Interview_On_Location_13_921.jpg', '33--Running/33_Running_Running_33_586.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_109.jpg', '52--Photographers/52_Photographers_taketouristphotos_52_208.jpg', '18--Concerts/18_Concerts_Concerts_18_670.jpg', '61--Street_Battle/61_Street_Battle_streetfight_61_546.jpg', '46--Jockey/46_Jockey_Jockey_46_779.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_405.jpg', '2--Demonstration/2_Demonstration_Protesters_2_577.jpg', '52--Photographers/52_Photographers_taketouristphotos_52_159.jpg', '54--Rescue/54_Rescue_firemanrescue_54_908.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_182.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_89.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_matadorbullfighting_47_645.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_matadorbullfighting_47_779.jpg', '8--Election_Campain/8_Election_Campain_Election_Campaign_8_433.jpg', '35--Basketball/35_Basketball_playingbasketball_35_651.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_688.jpg', '54--Rescue/54_Rescue_rescuepeople_54_924.jpg', '13--Interview/13_Interview_Interview_Sequences_13_936.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_380.jpg', '21--Festival/21_Festival_Festival_21_830.jpg', '15--Stock_Market/15_Stock_Market_Stock_Market_15_286.jpg', '12--Group/12_Group_Group_12_Group_Group_12_434.jpg', '12--Group/12_Group_Group_12_Group_Group_12_253.jpg', '3--Riot/3_Riot_Riot_3_186.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_655.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_607.jpg', '12--Group/12_Group_Group_12_Group_Group_12_198.jpg', '42--Car_Racing/42_Car_Racing_Car_Racing_42_263.jpg', '51--Dresses/51_Dresses_wearingdress_51_685.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_329.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_904.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_387.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_368.jpg', '2--Demonstration/2_Demonstration_Protesters_2_519.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_287.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_365.jpg', '37--Soccer/37_Soccer_soccer_ball_37_698.jpg', '39--Ice_Skating/39_Ice_Skating_iceskiing_39_272.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_456.jpg', '41--Swimming/41_Swimming_Swimmer_41_607.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_34.jpg', '30--Surgeons/30_Surgeons_Surgeons_30_256.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_Matador_Bullfighter_47_195.jpg', '51--Dresses/51_Dresses_wearingdress_51_183.jpg', '38--Tennis/38_Tennis_Tennis_38_692.jpg', '18--Concerts/18_Concerts_Concerts_18_313.jpg', '43--Row_Boat/43_Row_Boat_Canoe_43_784.jpg', '12--Group/12_Group_Group_12_Group_Group_12_301.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_249.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_639.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_567.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_640.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_192.jpg', '41--Swimming/41_Swimming_Swimmer_41_755.jpg', '12--Group/12_Group_Group_12_Group_Group_12_732.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_507.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_57.jpg', '43--Row_Boat/43_Row_Boat_Canoe_43_1048.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_374.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_750.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_18.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_64.jpg', '1--Handshaking/1_Handshaking_Handshaking_1_801.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_752.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_248.jpg', '59--people--driving--car/59_peopledrivingcar_peopledrivingcar_59_404.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_1026.jpg', '7--Cheering/7_Cheering_Cheering_7_134.jpg', '2--Demonstration/2_Demonstration_Protesters_2_351.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_115.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_793.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_213.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_483.jpg', '37--Soccer/37_Soccer_Soccer_37_655.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_395.jpg', '15--Stock_Market/15_Stock_Market_Stock_Market_15_460.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_644.jpg', '37--Soccer/37_Soccer_soccer_ball_37_1011.jpg', '52--Photographers/52_Photographers_photographertakingphoto_52_228.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_672.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_933.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_35.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_567.jpg', '13--Interview/13_Interview_Interview_On_Location_13_138.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_867.jpg', '35--Basketball/35_Basketball_Basketball_35_449.jpg', '0--Parade/0_Parade_marchingband_1_355.jpg', '48--Parachutist_Paratrooper/48_Parachutist_Paratrooper_Parachutist_Paratrooper_48_814.jpg', '53--Raid/53_Raid_policeraid_53_736.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_387.jpg', '53--Raid/53_Raid_policeraid_53_43.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_235.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_719.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_187.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_742.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_322.jpg', '56--Voter/56_Voter_peoplevoting_56_459.jpg', '2--Demonstration/2_Demonstration_Protesters_2_486.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_406.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_696.jpg', '13--Interview/13_Interview_Interview_Sequences_13_111.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_402.jpg', '61--Street_Battle/61_Street_Battle_streetfight_61_936.jpg', '3--Riot/3_Riot_Riot_3_710.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_684.jpg', '18--Concerts/18_Concerts_Concerts_18_60.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_232.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_103.jpg', '45--Balloonist/45_Balloonist_Balloonist_45_692.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_27.jpg', '15--Stock_Market/15_Stock_Market_Stock_Market_15_554.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_22.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_211.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_397.jpg', '36--Football/36_Football_americanfootball_ball_36_126.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_272.jpg', '37--Soccer/37_Soccer_soccer_ball_37_483.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_950.jpg', '18--Concerts/18_Concerts_Concerts_18_1038.jpg', '37--Soccer/37_Soccer_soccer_ball_37_692.jpg', '4--Dancing/4_Dancing_Dancing_4_715.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_850.jpg', '0--Parade/0_Parade_marchingband_1_1004.jpg', '38--Tennis/38_Tennis_Tennis_38_683.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_330.jpg', '21--Festival/21_Festival_Festival_21_225.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_403.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_162.jpg', '42--Car_Racing/42_Car_Racing_Car_Racing_42_602.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_891.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_19.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_1037.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_339.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_456.jpg', '22--Picnic/22_Picnic_Picnic_22_140.jpg', '12--Group/12_Group_Group_12_Group_Group_12_218.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_130.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_575.jpg', '28--Sports_Fan/28_Sports_Fan_Sports_Fan_28_590.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_931.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_60.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_597.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_17.jpg', '21--Festival/21_Festival_Festival_21_811.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_211.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_934.jpg', '3--Riot/3_Riot_Riot_3_273.jpg', '41--Swimming/41_Swimming_Swimming_41_822.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_716.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_807.jpg', '2--Demonstration/2_Demonstration_Protesters_2_174.jpg', '43--Row_Boat/43_Row_Boat_Canoe_43_940.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_10.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_430.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_712.jpg', '0--Parade/0_Parade_marchingband_1_139.jpg', '3--Riot/3_Riot_Riot_3_354.jpg', '35--Basketball/35_Basketball_Basketball_35_579.jpg', '52--Photographers/52_Photographers_photographertakingphoto_52_578.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_499.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_285.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_813.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_461.jpg', '38--Tennis/38_Tennis_Tennis_38_452.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_matadorbullfighting_47_300.jpg', '33--Running/33_Running_Running_33_547.jpg', '12--Group/12_Group_Group_12_Group_Group_12_293.jpg', '36--Football/36_Football_americanfootball_ball_36_257.jpg', '12--Group/12_Group_Group_12_Group_Group_12_59.jpg', '8--Election_Campain/8_Election_Campain_Election_Campaign_8_452.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_848.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_21.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_579.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_702.jpg', '54--Rescue/54_Rescue_rescuepeople_54_602.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_276.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_72.jpg', '2--Demonstration/2_Demonstration_Protesters_2_65.jpg', '35--Basketball/35_Basketball_playingbasketball_35_248.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_306.jpg', '2--Demonstration/2_Demonstration_Protesters_2_362.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_450.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_17.jpg', '42--Car_Racing/42_Car_Racing_Nascar_42_440.jpg', '7--Cheering/7_Cheering_Cheering_7_462.jpg', '6--Funeral/6_Funeral_Funeral_6_676.jpg', '16--Award_Ceremony/16_Award_Ceremony_Awards_Ceremony_16_569.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_546.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_matadorbullfighting_47_152.jpg', '18--Concerts/18_Concerts_Concerts_18_447.jpg', '2--Demonstration/2_Demonstration_Protesters_2_738.jpg', '0--Parade/0_Parade_marchingband_1_147.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_178.jpg', '2--Demonstration/2_Demonstration_Protesters_2_1033.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_252.jpg']
-test_labels = [[[445, 215, 143, 173]], [[452, 108, 208, 260]], [[372, 166, 328, 458]], [[293, 217, 179, 239]], [[672, 70, 114, 114]], [[370, 208, 186, 240]], [[436, 104, 134, 170]], [[556, 212, 198, 246]], [[362, 238, 350, 484]], [[306, 296, 148, 202]], [[433, 11, 447, 523]], [[297, 319, 17, 21]], [[520, 205, 475, 571]], [[754, 106, 106, 120]], [[154, 30, 78, 100]], [[200, 27, 379, 552]], [[281, 167, 399, 573]], [[433, 67, 177, 258]], [[183, 234, 366, 417]], [[490, 471, 125, 178]], [[770, 42, 84, 120]], [[694, 56, 180, 336]], [[419, 164, 336, 422]], [[318, 54, 250, 308]], [[352, 330, 364, 542]], [[294, 54, 456, 642]], [[614, 94, 110, 144]], [[492, 62, 98, 124]], [[344, 182, 236, 306]], [[448, 92, 140, 206], [540, 116, 142, 194]], [[350, 126, 52, 62], [674, 54, 56, 50]], [[186, 163, 86, 108], [446, 254, 35, 47]], [[790, 338, 88, 108], [128, 76, 70, 110]], [[192, 347, 17, 19], [485, 334, 22, 37]], [[448, 675, 157, 192], [651, 520, 165, 203]], [[346, 59, 49, 66], [453, 114, 46, 59]], [[360, 337, 295, 475], [610, 645, 315, 443]], [[284, 108, 94, 158], [472, 200, 122, 156]], [[235, 211, 315, 417], [615, 438, 185, 214]], [[542, 196, 25, 28], [914, 116, 46, 65]], [[270, 164, 196, 194], [664, 84, 84, 92]], [[331, 198, 10, 19], [428, 112, 12, 21]], [[683, 163, 29, 23], [322, 147, 7, 9]], [[358, 198, 172, 272], [588, 144, 160, 266]], [[196, 36, 128, 168], [670, 58, 100, 130]], [[434, 72, 256, 458], [214, 876, 78, 139]], [[335, 33, 99, 142], [630, 109, 92, 113]], [[498, 269, 31, 36], [196, 245, 35, 52]], [[92, 168, 270, 352], [500, 84, 218, 344]], [[306, 150, 256, 296], [860, 40, 150, 270]], [[229, 440, 133, 169], [392, 259, 232, 322]], [[416, 93, 144, 224], [6, 10, 112, 182]], [[440, 54, 84, 124], [536, 96, 90, 114]], [[301, 157, 214, 476], [536, 262, 386, 587]], [[332, 90, 64, 94], [770, 74, 60, 110]], [[502, 88, 92, 178], [636, 74, 110, 198]], [[697, 102, 41, 43], [943, 248, 32, 41]], [[190, 64, 82, 112], [720, 52, 72, 114]], [[396, 158, 236, 332], [680, 110, 256, 320], [86, 120, 216, 282]], [[766, 52, 54, 80], [612, 58, 56, 80], [270, 212, 60, 78]], [[857, 390, 13, 16], [622, 200, 47, 57], [414, 264, 30, 41]], [[300, 206, 122, 182], [638, 172, 132, 206], [476, 478, 106, 124]], [[613, 232, 24, 29], [498, 213, 24, 30], [419, 238, 24, 29]], [[336, 246, 96, 144], [745, 363, 117, 144], [622, 318, 69, 78]], [[684, 31, 43, 64], [511, 246, 35, 44], [864, 176, 40, 46]], [[334, 86, 214, 334], [678, 114, 200, 312], [38, 174, 186, 276]], [[679, 108, 40, 60], [513, 138, 49, 44], [427, 123, 36, 48]], [[287, 164, 51, 65], [555, 145, 47, 74], [894, 183, 58, 70]], [[509, 139, 7, 9], [938, 234, 7, 9], [508, 295, 7, 9]], [[370, 22, 64, 78], [518, 52, 60, 74], [714, 72, 78, 82]], [[270, 237, 50, 64], [506, 236, 59, 60], [756, 365, 71, 57]], [[214, 257, 27, 44], [384, 224, 32, 37], [228, 325, 34, 22]], [[268, 359, 84, 103], [685, 375, 81, 110], [475, 714, 110, 123]], [[214, 300, 96, 126], [458, 288, 86, 122], [708, 296, 80, 104]], [[220, 142, 116, 178], [422, 130, 136, 172], [650, 130, 120, 170]], [[644, 38, 62, 94], [440, 180, 74, 84], [32, 4, 62, 66]], [[308, 206, 78, 150], [442, 118, 98, 134], [552, 94, 88, 130]], [[562, 102, 62, 76], [646, 8, 74, 98], [60, 78, 124, 156]], [[200, 58, 48, 72], [396, 88, 53, 69], [941, 179, 6, 6]], [[336, 42, 80, 138], [560, 61, 86, 151], [243, 84, 55, 67]], [[387, 300, 13, 15], [185, 314, 13, 15], [714, 413, 15, 18]], [[408, 375, 14, 18], [483, 305, 10, 20], [421, 318, 11, 16]], [[780, 150, 42, 55], [798, 115, 36, 53], [349, 178, 49, 62]], [[668, 284, 53, 75], [411, 175, 67, 85], [255, 163, 66, 87]], [[245, 35, 40, 64], [492, 25, 43, 72], [743, 31, 40, 56]], [[448, 132, 88, 138], [662, 140, 86, 100], [946, 246, 74, 86]], [[154, 69, 130, 167], [690, 66, 111, 167], [594, 204, 64, 82]], [[189, 157, 26, 27], [407, 147, 22, 26], [499, 147, 19, 28], [800, 159, 24, 27]], [[96, 28, 72, 106], [350, 48, 66, 100], [606, 38, 68, 106], [866, 40, 80, 118]], [[276, 114, 60, 104], [404, 120, 54, 82], [570, 218, 60, 70], [776, 170, 52, 84]], [[490, 410, 15, 17], [552, 425, 16, 14], [618, 442, 17, 17], [680, 449, 16, 18]], [[776, 143, 49, 59], [564, 107, 68, 91], [392, 112, 38, 53], [124, 91, 63, 81]], [[180, 160, 72, 104], [424, 44, 78, 108], [750, 272, 66, 94], [872, 200, 106, 134]], [[200, 392, 21, 26], [660, 389, 26, 32], [753, 393, 27, 34], [935, 388, 18, 29]], [[78, 127, 23, 37], [375, 63, 77, 89], [600, 175, 17, 24], [329, 148, 24, 25]], [[179, 29, 60, 84], [291, 160, 63, 79], [404, 92, 64, 87], [450, 150, 50, 62]], [[242, 86, 136, 185], [672, 79, 37, 54], [820, 59, 38, 63], [960, 69, 40, 52]], [[810, 319, 28, 22], [296, 181, 30, 31], [691, 132, 11, 14], [785, 114, 11, 19]], [[99, 110, 14, 28], [56, 135, 21, 26], [427, 151, 24, 35], [529, 153, 19, 27]], [[935, 117, 41, 54], [537, 129, 51, 54], [994, 117, 29, 47], [790, 214, 32, 35]], [[140, 345, 26, 36], [400, 204, 44, 53], [550, 180, 54, 55], [853, 264, 25, 31]], [[810, 547, 27, 29], [608, 561, 19, 20], [477, 449, 18, 21], [161, 466, 21, 25]], [[288, 785, 36, 49], [438, 788, 39, 53], [584, 788, 35, 49], [747, 764, 37, 55]], [[438, 299, 70, 48], [619, 185, 23, 30], [858, 159, 28, 36], [852, 296, 36, 46]], [[610, 157, 48, 60], [379, 168, 49, 64], [626, 98, 26, 33], [819, 120, 34, 42]], [[132, 116, 73, 130], [403, 154, 80, 136], [661, 199, 75, 107], [784, 222, 53, 70]], [[672, 121, 49, 63], [593, 117, 51, 62], [527, 148, 47, 60], [391, 129, 51, 63]], [[803, 116, 37, 52], [787, 254, 37, 51], [784, 391, 34, 47], [151, 523, 8, 10]], [[839, 315, 49, 69], [640, 242, 66, 78], [519, 296, 59, 59], [169, 300, 66, 58]], [[716, 342, 90, 126], [638, 344, 82, 118], [176, 374, 76, 112], [236, 394, 116, 114]], [[272, 272, 221, 341], [722, 462, 79, 166], [794, 356, 127, 217], [562, 341, 66, 139]], [[306, 251, 31, 42], [473, 246, 30, 32], [637, 303, 36, 35], [842, 258, 21, 28]], [[349, 86, 20, 33], [478, 153, 29, 34], [574, 196, 22, 29], [668, 110, 20, 39]], [[26, 183, 31, 40], [461, 188, 54, 41], [570, 291, 20, 22], [522, 465, 6, 8]], [[477, 106, 307, 536], [823, 506, 54, 80], [927, 497, 52, 93], [4, 497, 60, 69]], [[689, 133, 92, 115], [623, 211, 59, 82], [488, 245, 40, 48], [397, 261, 28, 37]], [[550, 189, 31, 40], [813, 214, 44, 61], [408, 184, 44, 48], [158, 214, 35, 68], [171, 184, 41, 52]], [[142, 28, 98, 132], [290, 12, 86, 114], [472, 172, 88, 122], [580, 34, 80, 100], [774, 40, 100, 124]], [[208, 37, 67, 59], [791, 139, 20, 24], [388, 44, 43, 52], [549, 35, 40, 55], [613, 0, 37, 26]], [[664, 84, 56, 94], [858, 80, 72, 92], [302, 212, 60, 82], [104, 170, 80, 84], [326, 112, 58, 82]], [[335, 343, 25, 31], [372, 335, 22, 29], [570, 178, 19, 23], [536, 338, 21, 25], [616, 349, 23, 28]], [[174, 422, 68, 88], [351, 396, 80, 97], [491, 414, 66, 86], [667, 425, 63, 88], [847, 411, 68, 86]], [[30, 132, 67, 89], [400, 183, 42, 58], [606, 177, 28, 48], [801, 126, 42, 53], [958, 194, 43, 51]], [[896, 27, 51, 75], [661, 61, 58, 70], [428, 45, 55, 71], [283, 46, 49, 66], [93, 48, 53, 64]], [[326, 74, 16, 18], [874, 292, 15, 27], [943, 252, 13, 20], [682, 321, 20, 28], [680, 250, 11, 17]], [[262, 122, 51, 70], [321, 116, 58, 54], [449, 118, 55, 60], [530, 147, 65, 48], [551, 117, 52, 61]], [[283, 346, 118, 168], [491, 263, 100, 138], [463, 456, 108, 150], [709, 328, 108, 128], [588, 661, 108, 130]], [[344, 149, 83, 103], [435, 158, 25, 29], [576, 96, 106, 125], [707, 238, 81, 75], [135, 13, 22, 30]], [[549, 120, 59, 75], [796, 100, 51, 66], [898, 176, 39, 45], [161, 138, 65, 73], [364, 250, 34, 35]], [[565, 227, 47, 61], [730, 370, 37, 55], [314, 400, 43, 70], [731, 508, 21, 53], [3, 573, 42, 67]], [[469, 223, 63, 80], [261, 173, 49, 78], [659, 120, 67, 84], [795, 158, 61, 67], [796, 227, 52, 94]], [[682, 454, 50, 71], [889, 449, 49, 64], [499, 473, 69, 104], [296, 445, 126, 157], [43, 391, 78, 97]], [[142, 114, 118, 156], [352, 166, 100, 148], [572, 284, 82, 92], [720, 308, 96, 112], [902, 278, 120, 128]], [[94, 266, 118, 154], [98, 168, 92, 136], [236, 208, 126, 144], [382, 318, 110, 154], [854, 310, 122, 162]], [[226, 105, 5, 14], [395, 83, 122, 185], [673, 68, 7, 8], [832, 55, 8, 9], [517, 208, 15, 16]], [[525, 70, 17, 20], [609, 66, 16, 21], [691, 52, 19, 24], [762, 68, 23, 28], [95, 80, 12, 12]], [[670, 466, 116, 136], [690, 200, 166, 100], [436, 34, 122, 120], [268, 256, 150, 88], [322, 434, 146, 142]], [[288, 8, 51, 57], [895, 241, 68, 92], [756, 97, 62, 75], [702, 109, 49, 78], [570, 115, 54, 83]], [[74, 186, 102, 170], [282, 74, 112, 154], [466, 68, 86, 126], [614, 60, 110, 120], [876, 112, 120, 150]], [[775, 494, 20, 28], [482, 229, 48, 68], [227, 344, 48, 58], [745, 508, 11, 13], [641, 119, 292, 352]], [[126, 60, 92, 126], [312, 44, 96, 114], [488, 98, 186, 198], [748, 46, 80, 90], [912, 94, 100, 128]], [[206, 474, 28, 23], [222, 437, 24, 25], [349, 462, 29, 31], [400, 423, 20, 27], [657, 444, 23, 25]], [[130, 174, 126, 136], [346, 190, 84, 118], [474, 56, 54, 80], [594, 70, 62, 92], [726, 162, 60, 84]], [[149, 268, 61, 71], [353, 269, 53, 59], [572, 230, 53, 70], [659, 183, 55, 74], [806, 184, 59, 76]], [[34, 118, 112, 118], [320, 86, 74, 94], [548, 76, 66, 76], [722, 46, 58, 80], [908, 114, 72, 88]], [[63, 89, 56, 80], [237, 95, 49, 67], [431, 120, 43, 60], [537, 52, 65, 81], [700, 55, 52, 83], [902, 45, 49, 93]], [[354, 235, 164, 211], [619, 274, 86, 155], [750, 366, 80, 83], [831, 387, 71, 92], [217, 277, 89, 107], [6, 336, 95, 119]], [[70, 299, 32, 38], [91, 404, 27, 33], [516, 292, 35, 48], [574, 355, 32, 37], [637, 290, 36, 45], [563, 322, 20, 30]], [[879, 201, 16, 18], [863, 238, 16, 16], [871, 263, 17, 20], [939, 402, 20, 21], [872, 296, 15, 19], [908, 339, 18, 22]], [[466, 252, 30, 44], [588, 230, 27, 38], [698, 216, 31, 40], [822, 198, 26, 37], [341, 173, 31, 33], [222, 144, 33, 41]], [[852, 221, 59, 72], [724, 120, 61, 70], [411, 109, 38, 40], [462, 64, 36, 40], [293, 46, 20, 31], [343, 54, 24, 32]], [[729, 125, 40, 43], [909, 167, 39, 43], [577, 160, 37, 35], [424, 122, 37, 41], [286, 102, 41, 48], [127, 117, 41, 42]], [[209, 218, 57, 72], [334, 237, 56, 78], [437, 283, 53, 67], [567, 248, 51, 78], [667, 268, 52, 69], [760, 231, 58, 69]], [[690, 130, 29, 35], [675, 182, 20, 24], [471, 89, 11, 18], [391, 119, 22, 26], [416, 287, 51, 35], [452, 501, 40, 22]], [[0, 223, 128, 210], [140, 325, 10, 16], [161, 309, 12, 20], [323, 278, 47, 61], [456, 251, 102, 114], [768, 168, 58, 163]], [[229, 50, 55, 47], [587, 437, 12, 14], [506, 378, 16, 15], [830, 413, 14, 14], [751, 398, 19, 17], [905, 443, 16, 16]], [[858, 317, 70, 100], [776, 242, 32, 51], [654, 273, 56, 76], [535, 239, 41, 54], [244, 268, 41, 54], [109, 318, 54, 67]], [[869, 27, 98, 133], [645, 44, 54, 70], [556, 79, 49, 67], [410, 23, 58, 69], [243, 66, 58, 68], [69, 60, 62, 69]], [[810, 196, 182, 269], [644, 169, 181, 234], [488, 153, 114, 151], [340, 138, 90, 129], [15, 398, 48, 63], [98, 392, 27, 43]], [[495, 191, 36, 39], [564, 67, 21, 29], [396, 240, 43, 58], [85, 48, 16, 19], [689, 20, 15, 15], [824, 18, 15, 16]], [[865, 182, 49, 65], [794, 116, 16, 18], [741, 82, 14, 17], [696, 85, 16, 17], [660, 104, 10, 11], [200, 395, 89, 84]], [[220, 235, 75, 75], [368, 285, 60, 88], [591, 250, 55, 68], [678, 233, 73, 115], [473, 483, 65, 88], [203, 453, 63, 85]], [[6, 34, 68, 144], [162, 72, 78, 118], [486, 146, 60, 80], [248, 222, 90, 148], [472, 236, 86, 124], [724, 256, 84, 112]], [[105, 25, 51, 56], [226, 118, 33, 37], [389, 0, 38, 50], [422, 52, 25, 54], [602, 493, 58, 32], [613, 52, 31, 44]], [[166, 144, 74, 102], [326, 246, 56, 80], [376, 146, 72, 96], [576, 168, 66, 86], [674, 118, 82, 84], [780, 82, 96, 116]], [[786, 371, 17, 28], [511, 554, 296, 341], [844, 985, 26, 36], [146, 642, 101, 115], [571, 1076, 40, 36], [945, 831, 19, 18]], [[114, 158, 96, 168], [200, 324, 102, 134], [416, 272, 114, 140], [596, 220, 96, 158], [696, 198, 102, 176], [938, 438, 66, 152]], [[121, 427, 77, 98], [204, 318, 65, 92], [348, 154, 81, 119], [687, 237, 75, 106], [812, 250, 83, 98], [512, 156, 81, 114]], [[23, 84, 28, 43], [358, 16, 34, 55], [485, 30, 26, 44], [615, 126, 19, 32], [736, 125, 26, 40], [572, 186, 10, 13]], [[553, 340, 53, 87], [750, 338, 51, 78], [872, 338, 61, 74], [804, 478, 136, 132], [981, 277, 43, 105], [954, 246, 29, 45]], [[775, 73, 62, 73], [579, 112, 51, 73], [449, 143, 52, 64], [334, 120, 56, 74], [213, 43, 57, 76], [145, 127, 60, 74]], [[552, 275, 21, 38], [396, 188, 20, 35], [274, 171, 30, 38], [207, 153, 20, 33], [54, 98, 21, 28], [126, 110, 30, 31]], [[840, 279, 28, 35], [776, 320, 38, 52], [481, 1, 297, 456], [324, 0, 100, 134], [401, 227, 79, 106], [430, 0, 44, 63]], [[473, 143, 31, 42], [531, 191, 32, 37], [594, 161, 30, 43], [346, 152, 33, 46], [389, 168, 30, 40], [774, 82, 33, 66]], [[73, 175, 27, 39], [224, 165, 31, 44], [267, 168, 30, 38], [140, 242, 58, 60], [453, 172, 27, 35], [581, 217, 60, 79], [644, 181, 137, 169]], [[110, 198, 38, 51], [539, 159, 27, 30], [521, 182, 25, 27], [985, 145, 18, 27], [771, 189, 14, 26], [320, 72, 13, 23], [238, 51, 16, 20]], [[463, 687, 144, 199], [714, 527, 95, 132], [843, 331, 77, 98], [561, 319, 71, 80], [104, 362, 74, 92], [120, 233, 52, 83], [509, 251, 67, 86]], [[15, 279, 15, 20], [171, 242, 14, 20], [355, 292, 95, 87], [445, 135, 40, 71], [549, 108, 119, 99], [907, 135, 42, 51], [511, 157, 36, 47]], [[47, 334, 41, 52], [98, 262, 16, 21], [335, 228, 15, 21], [414, 247, 9, 14], [505, 250, 17, 20], [671, 278, 18, 22], [381, 264, 6, 9]], [[808, 254, 17, 26], [716, 284, 22, 32], [561, 287, 24, 31], [508, 264, 31, 44], [390, 259, 32, 35], [232, 280, 24, 29], [60, 298, 28, 34]], [[79, 409, 11, 15], [288, 388, 12, 17], [493, 243, 22, 33], [563, 425, 8, 11], [855, 397, 11, 15], [905, 416, 9, 11], [982, 409, 8, 12]], [[355, 104, 22, 28], [456, 114, 26, 33], [562, 90, 31, 39], [689, 107, 42, 50], [816, 177, 29, 39], [888, 107, 25, 36], [905, 87, 21, 27]], [[57, 412, 24, 25], [156, 409, 26, 28], [211, 411, 22, 24], [251, 402, 24, 28], [307, 400, 18, 23], [378, 209, 231, 279], [695, 351, 207, 340]], [[111, 221, 31, 35], [188, 223, 32, 35], [284, 201, 35, 39], [490, 185, 43, 51], [376, 186, 37, 43], [619, 316, 39, 46], [836, 337, 42, 50]], [[690, 246, 31, 38], [636, 243, 28, 41], [516, 225, 32, 43], [253, 279, 16, 18], [408, 294, 5, 6], [375, 289, 3, 5], [564, 261, 9, 15]], [[88, 88, 40, 46], [261, 122, 40, 49], [163, 185, 44, 34], [470, 118, 35, 44], [618, 129, 40, 55], [654, 39, 30, 39], [771, 88, 45, 55]], [[424, 566, 62, 83], [75, 57, 173, 256], [207, 393, 155, 217], [543, 34, 349, 432], [833, 494, 124, 165], [750, 990, 163, 225], [91, 988, 178, 246]], [[461, 214, 28, 40], [63, 239, 22, 25], [126, 250, 20, 27], [205, 148, 23, 29], [383, 250, 18, 20], [816, 243, 27, 29], [760, 298, 32, 32]], [[387, 23, 378, 637], [670, 674, 44, 55], [588, 741, 41, 52], [60, 879, 36, 44], [215, 1255, 84, 89], [638, 1175, 41, 54], [224, 1404, 44, 79]], [[776, 258, 24, 38], [702, 242, 29, 31], [594, 259, 24, 32], [463, 135, 34, 54], [341, 218, 22, 35], [273, 244, 27, 43], [164, 194, 26, 35]], [[61, 349, 29, 37], [252, 352, 35, 50], [320, 366, 28, 37], [516, 373, 36, 51], [757, 364, 28, 40], [817, 396, 32, 48], [999, 380, 25, 39]], [[742, 89, 35, 47], [360, 72, 25, 47], [246, 75, 31, 65], [178, 147, 26, 32], [498, 80, 19, 39], [966, 155, 24, 33], [860, 90, 29, 42]], [[446, 133, 163, 162], [815, 115, 118, 118], [45, 400, 133, 154], [232, 425, 55, 60], [24, 542, 39, 39], [597, 239, 80, 80], [802, 306, 35, 49]], [[531, 215, 43, 62], [363, 219, 58, 77], [251, 262, 69, 92], [116, 252, 80, 99], [707, 248, 79, 90], [870, 250, 90, 126], [280, 124, 28, 31]], [[111, 391, 83, 108], [270, 283, 76, 127], [369, 410, 99, 105], [483, 324, 73, 111], [642, 308, 73, 111], [798, 328, 80, 105], [560, 572, 76, 89]], [[122, 211, 23, 31], [217, 150, 31, 48], [290, 170, 27, 35], [389, 220, 22, 26], [508, 176, 25, 34], [596, 33, 41, 53], [763, 117, 30, 42]], [[188, 158, 54, 54], [316, 160, 56, 56], [472, 130, 52, 46], [392, 114, 54, 60], [606, 116, 58, 54], [668, 144, 64, 54], [798, 122, 60, 62]], [[496, 218, 38, 54], [671, 325, 30, 40], [706, 327, 32, 37], [124, 265, 41, 50], [349, 333, 41, 39], [832, 380, 34, 42], [963, 259, 41, 50]], [[491, 177, 56, 64], [616, 71, 59, 73], [29, 184, 28, 32], [0, 193, 26, 35], [292, 82, 37, 54], [743, 12, 33, 66], [153, 199, 16, 19]], [[456, 42, 47, 62], [211, 66, 49, 55], [787, 478, 7, 9], [988, 474, 7, 6], [955, 457, 6, 7], [294, 512, 7, 9], [322, 520, 4, 6]], [[718, 403, 15, 27], [707, 477, 26, 30], [502, 397, 23, 41], [256, 379, 19, 28], [93, 361, 18, 25], [806, 536, 9, 11], [975, 621, 15, 17]], [[153, 191, 14, 35], [482, 87, 21, 35], [696, 187, 7, 9], [771, 191, 6, 13], [826, 199, 11, 23], [644, 183, 5, 11], [781, 196, 4, 9]], [[719, 239, 26, 38], [415, 187, 30, 42], [443, 169, 22, 27], [336, 168, 17, 20], [181, 183, 16, 23], [227, 213, 17, 20], [533, 179, 13, 23]], [[814, 734, 36, 34], [683, 741, 10, 13], [782, 717, 8, 11], [657, 716, 7, 10], [701, 720, 12, 15], [906, 724, 7, 12], [948, 721, 11, 15], [933, 716, 8, 10]], [[909, 335, 65, 84], [786, 187, 76, 104], [718, 230, 59, 82], [595, 315, 72, 96], [342, 392, 53, 82], [302, 152, 66, 96], [178, 136, 79, 108], [522, 41, 58, 72]], [[485, 200, 104, 114], [217, 73, 157, 208], [339, 109, 96, 132], [0, 175, 69, 94], [678, 217, 38, 52], [667, 150, 13, 14], [114, 182, 38, 51], [738, 255, 23, 27]], [[415, 288, 38, 53], [350, 293, 24, 44], [620, 252, 42, 67], [340, 254, 30, 53], [26, 113, 35, 55], [241, 242, 28, 56], [728, 249, 23, 54], [796, 294, 37, 62]], [[64, 149, 51, 61], [208, 162, 47, 59], [342, 174, 41, 55], [446, 194, 44, 59], [670, 184, 44, 59], [773, 183, 45, 59], [920, 216, 45, 53], [291, 108, 27, 45]], [[87, 90, 47, 63], [208, 134, 46, 52], [355, 103, 45, 61], [483, 118, 41, 55], [627, 111, 45, 61], [712, 153, 40, 52], [783, 161, 45, 53], [869, 142, 53, 61]], [[60, 139, 31, 48], [340, 164, 31, 46], [782, 115, 32, 48], [523, 300, 33, 46], [329, 304, 40, 39], [675, 94, 35, 43], [699, 140, 42, 44], [641, 224, 42, 42]], [[116, 41, 47, 56], [276, 30, 49, 59], [384, 93, 42, 58], [481, 53, 42, 63], [548, 37, 48, 65], [654, 86, 43, 58], [727, 26, 41, 62], [832, 47, 44, 56]], [[98, 152, 40, 57], [205, 167, 31, 39], [495, 88, 46, 60], [401, 120, 44, 44], [684, 146, 36, 45], [761, 144, 27, 36], [853, 232, 13, 17], [912, 80, 37, 41]], [[108, 85, 69, 71], [339, 48, 65, 76], [486, 79, 55, 68], [643, 81, 61, 74], [862, 52, 67, 84], [957, 24, 33, 38], [1009, 30, 15, 33], [978, 9, 21, 29]], [[626, 142, 82, 102], [536, 88, 76, 88], [450, 156, 84, 110], [276, 162, 86, 118], [288, 632, 100, 148], [90, 660, 102, 158], [462, 606, 100, 124], [632, 572, 94, 108]], [[460, 266, 9, 11], [578, 264, 10, 13], [571, 258, 8, 10], [618, 261, 10, 14], [658, 262, 9, 13], [80, 272, 10, 13], [74, 256, 8, 10], [459, 210, 6, 7]], [[411, 109, 41, 61], [658, 123, 43, 60], [488, 469, 38, 46], [417, 445, 30, 34], [295, 467, 34, 36], [146, 454, 33, 37], [74, 457, 39, 43], [626, 452, 34, 42]], [[471, 167, 128, 161], [269, 234, 137, 142], [670, 266, 33, 47], [0, 93, 50, 209], [929, 283, 8, 9], [966, 293, 10, 10], [913, 278, 7, 11], [1011, 276, 5, 10]], [[119, 210, 38, 46], [155, 406, 49, 55], [330, 406, 46, 43], [429, 229, 37, 47], [648, 188, 42, 52], [688, 204, 38, 43], [771, 205, 35, 47], [811, 229, 35, 34]], [[797, 154, 54, 69], [182, 156, 49, 68], [337, 150, 58, 80], [429, 151, 51, 67], [467, 221, 57, 75], [590, 154, 51, 78], [647, 179, 61, 77], [397, 429, 7, 9]], [[96, 63, 75, 97], [286, 98, 65, 90], [384, 101, 60, 77], [465, 78, 56, 75], [578, 72, 63, 70], [851, 76, 67, 89], [808, 46, 59, 74], [694, 127, 59, 73]], [[383, 446, 46, 62], [667, 215, 46, 60], [621, 466, 54, 67], [747, 281, 48, 63], [844, 307, 55, 66], [822, 281, 7, 10], [965, 232, 9, 14], [423, 233, 9, 11]], [[436, 334, 22, 28], [359, 410, 22, 28], [317, 333, 22, 25], [232, 396, 22, 30], [185, 290, 21, 27], [603, 263, 20, 25], [752, 264, 21, 27], [881, 212, 22, 27]], [[146, 98, 64, 92], [244, 130, 58, 92], [310, 140, 60, 86], [456, 110, 60, 86], [514, 140, 64, 84], [672, 130, 54, 72], [768, 120, 56, 76], [846, 104, 58, 88]], [[959, 231, 13, 18], [823, 173, 8, 9], [813, 181, 9, 9], [85, 176, 7, 10], [0, 258, 10, 29], [136, 178, 5, 6], [774, 183, 7, 10], [302, 173, 3, 8]], [[488, 23, 164, 205], [306, 270, 136, 183], [181, 385, 47, 56], [233, 291, 32, 45], [31, 390, 50, 53], [31, 263, 22, 37], [58, 265, 16, 20], [158, 259, 12, 17]], [[133, 373, 44, 53], [212, 425, 27, 41], [518, 395, 40, 50], [453, 419, 24, 33], [405, 435, 30, 40], [862, 428, 44, 60], [654, 445, 29, 44], [747, 491, 15, 21]], [[0, 170, 83, 145], [140, 121, 76, 118], [219, 43, 73, 112], [343, 45, 70, 100], [568, 48, 64, 84], [759, 104, 67, 77], [968, 108, 56, 129], [913, 137, 72, 98]], [[615, 121, 28, 45], [456, 119, 29, 37], [278, 137, 32, 49], [165, 149, 41, 53], [137, 211, 33, 58], [675, 146, 34, 59], [720, 185, 37, 57], [894, 253, 90, 119]], [[2, 286, 112, 108], [168, 208, 92, 118], [342, 30, 110, 144], [846, 222, 100, 136], [460, 356, 78, 84], [634, 414, 80, 116], [904, 480, 78, 97], [208, 456, 72, 78]], [[316, 345, 21, 29], [418, 350, 26, 29], [569, 371, 21, 27], [786, 387, 24, 29], [921, 377, 17, 24], [868, 379, 19, 24], [922, 429, 14, 27], [635, 390, 23, 22]], [[850, 139, 51, 94], [850, 62, 52, 68], [709, 66, 49, 57], [460, 92, 52, 42], [221, 69, 56, 54], [78, 48, 58, 58], [111, 231, 65, 50], [678, 176, 48, 93]], [[106, 268, 48, 56], [378, 165, 38, 56], [528, 244, 37, 40], [239, 245, 33, 32], [324, 207, 12, 19], [681, 492, 37, 54], [614, 420, 43, 66], [470, 199, 25, 38]], [[195, 233, 57, 67], [277, 184, 61, 70], [342, 273, 57, 69], [430, 206, 58, 64], [489, 297, 56, 70], [529, 137, 56, 57], [624, 247, 62, 71], [694, 196, 52, 61], [770, 264, 58, 63]], [[47, 39, 30, 36], [30, 270, 29, 35], [199, 293, 26, 45], [238, 249, 30, 35], [549, 171, 29, 36], [549, 308, 30, 33], [598, 613, 34, 34], [771, 462, 32, 38], [916, 216, 30, 32]], [[497, 96, 71, 87], [638, 161, 66, 78], [364, 65, 25, 29], [438, 52, 19, 25], [258, 49, 21, 26], [733, 56, 19, 25], [915, 96, 11, 15], [877, 110, 8, 10], [841, 112, 8, 9]], [[147, 504, 23, 43], [392, 579, 15, 25], [695, 607, 5, 9], [718, 619, 6, 8], [708, 609, 8, 9], [775, 611, 7, 13], [583, 572, 10, 15], [666, 614, 6, 9], [665, 633, 5, 7]], [[783, 286, 40, 45], [963, 319, 29, 32], [725, 330, 22, 31], [567, 316, 21, 31], [416, 319, 43, 55], [326, 344, 24, 30], [154, 344, 13, 15], [106, 347, 19, 28], [1, 316, 51, 47]], [[828, 346, 51, 61], [596, 298, 52, 41], [477, 246, 47, 30], [486, 123, 36, 36], [378, 65, 27, 38], [381, 253, 47, 40], [207, 189, 37, 38], [159, 147, 35, 38], [275, 319, 53, 63]], [[877, 271, 32, 30], [751, 258, 28, 32], [595, 266, 34, 32], [233, 400, 32, 22], [95, 351, 35, 36], [559, 363, 13, 19], [662, 366, 18, 16], [827, 333, 11, 16], [488, 274, 30, 30]], [[586, 370, 44, 54], [838, 157, 54, 63], [769, 206, 42, 59], [685, 223, 42, 49], [592, 238, 38, 48], [553, 233, 34, 45], [218, 210, 46, 52], [135, 203, 29, 61], [409, 103, 30, 68]], [[1, 169, 23, 63], [68, 164, 43, 60], [146, 127, 34, 61], [145, 389, 81, 109], [478, 59, 93, 127], [679, 214, 73, 82], [731, 172, 49, 77], [972, 212, 49, 60], [802, 551, 97, 117]], [[174, 97, 45, 58], [400, 99, 39, 52], [306, 235, 26, 35], [238, 323, 28, 34], [349, 336, 20, 30], [379, 323, 16, 22], [103, 419, 37, 53], [462, 384, 45, 55], [65, 173, 19, 22]], [[581, 457, 15, 17], [643, 435, 14, 20], [727, 448, 11, 15], [730, 525, 15, 19], [531, 387, 6, 16], [873, 562, 20, 26], [193, 738, 32, 30], [299, 698, 30, 43], [882, 745, 22, 23]], [[2, 433, 34, 45], [75, 435, 38, 50], [122, 419, 40, 41], [196, 407, 45, 50], [326, 345, 51, 67], [507, 290, 62, 77], [767, 271, 76, 85], [653, 524, 19, 24], [930, 454, 32, 36]], [[626, 72, 42, 57], [555, 28, 41, 54], [409, 77, 50, 55], [892, 239, 66, 104], [970, 187, 53, 65], [275, 177, 65, 36], [138, 188, 41, 77], [687, 414, 79, 118], [777, 96, 47, 69]], [[657, 47, 118, 157], [509, 183, 57, 70], [991, 202, 27, 27], [441, 279, 35, 40], [42, 370, 40, 48], [18, 441, 25, 31], [0, 461, 28, 33], [235, 256, 65, 69], [585, 248, 70, 78]], [[120, 398, 25, 34], [0, 278, 9, 32], [247, 239, 25, 38], [310, 190, 29, 33], [398, 201, 28, 43], [515, 212, 25, 33], [677, 297, 27, 40], [757, 296, 29, 41], [529, 427, 30, 41]], [[62, 383, 40, 53], [275, 308, 37, 47], [469, 332, 25, 35], [586, 310, 19, 43], [639, 296, 24, 37], [1003, 334, 14, 24], [165, 342, 15, 30], [535, 311, 17, 19], [761, 329, 17, 24]], [[163, 162, 35, 40], [63, 130, 22, 18], [213, 156, 26, 30], [88, 144, 20, 23], [287, 129, 31, 36], [416, 171, 30, 37], [631, 200, 46, 53], [768, 196, 41, 50], [913, 251, 35, 42]], [[641, 530, 43, 62], [701, 536, 24, 27], [682, 529, 17, 31], [761, 467, 44, 60], [889, 431, 36, 41], [989, 368, 35, 61], [812, 514, 51, 52], [175, 601, 34, 51], [117, 566, 25, 35]], [[693, 257, 171, 101], [839, 81, 82, 90], [303, 46, 32, 40], [224, 84, 38, 45], [392, 84, 34, 56], [456, 93, 33, 52], [945, 202, 34, 47], [72, 80, 40, 48], [720, 125, 52, 54]], [[393, 240, 23, 54], [406, 161, 23, 32], [498, 176, 22, 34], [239, 151, 19, 27], [536, 103, 12, 16], [620, 99, 14, 18], [701, 104, 13, 21], [565, 138, 11, 24], [916, 109, 18, 23]], [[903, 96, 51, 64], [492, 116, 23, 36], [733, 299, 47, 54], [204, 216, 23, 25], [80, 96, 28, 29], [411, 343, 24, 27], [478, 572, 22, 28], [266, 529, 23, 31], [77, 599, 25, 29]], [[66, 144, 80, 110], [110, 110, 74, 106], [276, 164, 76, 100], [352, 124, 56, 80], [498, 250, 64, 82], [572, 140, 56, 86], [758, 170, 52, 74], [806, 138, 54, 90], [948, 168, 62, 70]], [[558, 288, 23, 25], [491, 272, 22, 30], [444, 275, 20, 25], [377, 245, 23, 29], [737, 82, 25, 28], [639, 75, 25, 24], [600, 73, 17, 20], [539, 29, 19, 24], [847, 0, 18, 10]], [[960, 275, 45, 40], [784, 83, 14, 17], [568, 169, 34, 49], [567, 93, 21, 26], [431, 135, 28, 33], [288, 153, 33, 33], [318, 94, 8, 9], [41, 135, 22, 19], [306, 141, 25, 34]], [[261, 330, 16, 37], [333, 345, 21, 27], [350, 362, 15, 25], [536, 351, 18, 30], [700, 390, 16, 18], [656, 450, 16, 17], [713, 478, 14, 14], [834, 370, 21, 31], [472, 366, 15, 21]], [[650, 187, 17, 24], [688, 141, 18, 26], [744, 145, 18, 31], [604, 174, 12, 23], [759, 80, 7, 12], [772, 85, 9, 10], [448, 185, 8, 14], [385, 158, 9, 14], [918, 143, 12, 19]], [[46, 114, 37, 51], [185, 127, 39, 52], [282, 79, 37, 51], [393, 63, 33, 51], [510, 55, 34, 52], [614, 60, 38, 49], [722, 79, 36, 50], [807, 99, 38, 51], [913, 112, 33, 52]], [[926, 363, 31, 38], [818, 341, 32, 32], [839, 96, 27, 36], [225, 541, 33, 35], [101, 469, 25, 38], [9, 0, 20, 30], [0, 94, 13, 36], [16, 156, 17, 32], [55, 255, 15, 35]], [[117, 44, 57, 54], [255, 87, 49, 47], [363, 107, 48, 49], [452, 54, 51, 52], [523, 112, 51, 58], [622, 141, 47, 48], [743, 69, 55, 56], [938, 177, 13, 14], [7, 138, 5, 7]], [[905, 143, 43, 49], [754, 150, 41, 59], [820, 154, 41, 56], [648, 90, 30, 31], [519, 144, 57, 62], [571, 124, 35, 56], [592, 133, 28, 50], [368, 107, 48, 57], [5, 84, 32, 36], [106, 147, 59, 49]], [[71, 222, 75, 95], [429, 193, 63, 81], [292, 179, 49, 65], [100, 166, 55, 65], [629, 132, 54, 72], [864, 152, 83, 123], [503, 205, 35, 53], [387, 191, 29, 48], [780, 209, 44, 61], [62, 185, 19, 27]], [[112, 87, 27, 32], [172, 106, 27, 41], [275, 86, 19, 24], [216, 195, 31, 33], [240, 237, 31, 65], [745, 172, 23, 35], [806, 145, 26, 40], [523, 89, 19, 29], [592, 121, 25, 35], [415, 84, 19, 24]], [[0, 414, 31, 57], [65, 445, 39, 46], [77, 449, 46, 53], [152, 450, 48, 48], [240, 418, 60, 59], [369, 400, 50, 66], [421, 389, 67, 74], [645, 293, 69, 83], [788, 333, 91, 91], [969, 389, 55, 72]], [[131, 187, 33, 43], [243, 214, 33, 47], [363, 204, 39, 43], [481, 156, 44, 59], [584, 200, 24, 39], [641, 165, 50, 58], [728, 177, 33, 47], [760, 128, 54, 73], [989, 188, 33, 58], [105, 243, 26, 29]], [[563, 305, 25, 26], [356, 299, 25, 31], [471, 277, 22, 26], [521, 313, 17, 25], [264, 320, 38, 45], [422, 314, 9, 13], [427, 297, 11, 14], [413, 307, 6, 9], [498, 303, 13, 16], [314, 323, 7, 8]], [[142, 166, 68, 92], [338, 172, 54, 88], [454, 178, 58, 88], [556, 204, 74, 100], [650, 128, 60, 102], [776, 130, 64, 86], [862, 186, 74, 104], [236, 452, 68, 88], [488, 412, 58, 98], [726, 398, 66, 86]], [[831, 83, 41, 46], [639, 115, 20, 26], [385, 169, 16, 21], [417, 169, 12, 21], [346, 127, 7, 7], [373, 148, 7, 8], [175, 58, 47, 56], [401, 174, 12, 17], [349, 165, 8, 9], [472, 170, 9, 10]], [[218, 197, 59, 72], [351, 229, 53, 75], [499, 200, 59, 85], [647, 229, 60, 74], [839, 240, 78, 79], [801, 1444, 5, 7], [811, 1445, 6, 9], [843, 1440, 9, 10], [863, 1443, 7, 9], [947, 1447, 8, 9]], [[17, 0, 58, 53], [236, 0, 56, 48], [18, 125, 40, 65], [67, 184, 58, 62], [277, 160, 56, 81], [492, 119, 55, 77], [432, 50, 48, 45], [671, 120, 66, 89], [961, 34, 52, 75], [940, 151, 61, 76]], [[287, 104, 48, 60], [216, 259, 57, 72], [408, 179, 49, 69], [454, 350, 44, 52], [540, 467, 50, 60], [355, 466, 56, 73], [494, 217, 65, 86], [616, 233, 47, 59], [673, 135, 53, 69], [711, 373, 57, 75]], [[930, 317, 44, 63], [734, 215, 74, 103], [560, 254, 62, 83], [431, 313, 29, 52], [320, 294, 34, 46], [275, 315, 37, 49], [272, 279, 32, 37], [161, 270, 40, 55], [125, 300, 33, 41], [96, 327, 22, 29]], [[172, 894, 24, 25], [286, 857, 12, 13], [145, 859, 16, 19], [154, 959, 30, 35], [396, 833, 31, 40], [493, 930, 29, 32], [651, 864, 20, 20], [816, 863, 19, 17], [624, 905, 41, 50], [27, 895, 31, 38]], [[0, 260, 62, 107], [175, 167, 123, 148], [318, 204, 37, 46], [362, 183, 57, 70], [457, 218, 40, 28], [590, 236, 45, 65], [663, 213, 63, 86], [748, 261, 37, 64], [826, 226, 101, 127], [617, 204, 48, 59]], [[808, 412, 77, 77], [867, 538, 20, 30], [885, 523, 22, 29], [385, 523, 12, 16], [368, 446, 5, 11], [417, 498, 13, 13], [309, 330, 23, 27], [375, 471, 6, 10], [448, 507, 13, 15], [710, 556, 17, 26]], [[573, 197, 5, 8], [589, 197, 4, 6], [595, 200, 5, 7], [663, 357, 15, 17], [456, 333, 6, 15], [431, 680, 18, 24], [855, 633, 18, 28], [784, 591, 23, 30], [247, 433, 19, 25], [344, 268, 7, 8]], [[68, 93, 63, 70], [446, 488, 25, 35], [165, 499, 45, 46], [277, 454, 21, 25], [542, 449, 20, 25], [375, 238, 18, 22], [309, 360, 25, 17], [193, 464, 15, 14], [856, 63, 53, 70], [17, 309, 92, 119]], [[0, 171, 12, 21], [66, 160, 17, 20], [156, 171, 12, 19], [166, 170, 13, 20], [599, 175, 23, 25], [657, 139, 23, 24], [696, 163, 23, 28], [730, 149, 18, 33], [804, 158, 14, 20], [866, 169, 13, 17]], [[819, 362, 36, 51], [197, 369, 30, 36], [273, 365, 39, 38], [59, 421, 29, 36], [550, 132, 67, 97], [0, 545, 18, 40], [68, 553, 24, 49], [361, 589, 31, 71], [441, 592, 31, 59], [721, 336, 25, 43]], [[616, 258, 51, 76], [497, 250, 36, 49], [558, 144, 53, 69], [318, 273, 42, 29], [732, 37, 28, 40], [853, 45, 34, 42], [906, 237, 47, 59], [662, 182, 42, 50], [645, 57, 32, 51], [254, 390, 57, 44]], [[0, 31, 39, 64], [88, 23, 71, 86], [163, 88, 28, 54], [191, 35, 41, 56], [261, 16, 69, 93], [336, 81, 59, 57], [445, 9, 65, 77], [564, 149, 141, 170], [772, 38, 120, 191], [830, 98, 117, 197]], [[38, 280, 8, 10], [83, 285, 7, 12], [216, 248, 64, 72], [296, 295, 9, 14], [387, 285, 54, 70], [466, 299, 8, 9], [499, 297, 6, 11], [534, 282, 14, 21], [574, 288, 13, 25], [645, 316, 19, 28]], [[778, 274, 28, 48], [994, 27, 19, 27], [683, 44, 25, 29], [626, 3, 24, 30], [194, 213, 30, 46], [469, 0, 20, 21], [424, 0, 18, 20], [310, 7, 20, 25], [171, 9, 19, 23], [117, 11, 16, 19]], [[353, 483, 37, 45], [344, 368, 45, 48], [212, 385, 38, 40], [246, 524, 35, 39], [119, 547, 40, 39], [88, 379, 35, 48], [399, 189, 41, 42], [244, 194, 33, 23], [70, 196, 48, 55], [901, 120, 54, 55]], [[219, 33, 79, 97], [309, 59, 55, 75], [474, 116, 91, 112], [766, 170, 70, 100], [809, 81, 44, 47], [32, 1224, 81, 204], [250, 884, 114, 149], [725, 1028, 119, 150], [925, 1063, 99, 156], [672, 698, 42, 32]], [[506, 89, 80, 90], [304, 126, 56, 73], [400, 170, 66, 71], [872, 153, 38, 56], [216, 154, 48, 59], [159, 123, 43, 47], [90, 146, 32, 39], [618, 144, 30, 40], [17, 168, 22, 25], [317, 38, 14, 19]], [[876, 110, 108, 154], [702, 46, 84, 136], [554, 86, 88, 120], [450, 96, 90, 110], [344, 140, 80, 92], [290, 118, 68, 96], [186, 122, 68, 74], [122, 76, 70, 86], [640, 102, 74, 96], [756, 94, 72, 112]], [[834, 424, 13, 19], [807, 429, 10, 12], [783, 429, 9, 12], [562, 427, 8, 10], [548, 416, 12, 13], [495, 344, 25, 38], [419, 314, 26, 41], [362, 266, 28, 53], [317, 194, 49, 87], [108, 51, 90, 102]], [[186, 190, 41, 41], [307, 179, 18, 18], [323, 175, 23, 28], [376, 182, 22, 28], [417, 167, 20, 29], [474, 183, 26, 30], [514, 177, 25, 28], [627, 197, 18, 20], [719, 181, 31, 38], [822, 164, 23, 30]]]
