--- val_data.py
@@ -1,2 +1,0 @@
-val_data = ['41--Swimming/41_Swimming_Swimmer_41_449.jpg', '27--Spa/27_Spa_Spa_27_225.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_769.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_267.jpg', '6--Funeral/6_Funeral_Funeral_6_690.jpg', '43--Row_Boat/43_Row_Boat_Canoe_43_881.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_364.jpg', '56--Voter/56_Voter_peoplevoting_56_1046.jpg', '41--Swimming/41_Swimming_Swimmer_41_610.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_203.jpg', '33--Running/33_Running_Running_33_209.jpg', '43--Row_Boat/43_Row_Boat_Rowboat_43_688.jpg', '28--Sports_Fan/28_Sports_Fan_Sports_Fan_28_86.jpg', '18--Concerts/18_Concerts_Concerts_18_528.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_870.jpg', '4--Dancing/4_Dancing_Dancing_4_885.jpg', '13--Interview/13_Interview_Interview_Sequences_13_33.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_368.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_155.jpg', '27--Spa/27_Spa_Spa_27_212.jpg', '52--Photographers/52_Photographers_photographertakingphoto_52_125.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_431.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_68.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_Matador_Bullfighter_47_631.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_847.jpg', '59--people--driving--car/59_peopledrivingcar_peopledrivingcar_59_592.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_683.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_Matador_Bullfighter_47_703.jpg', '54--Rescue/54_Rescue_firemanrescue_54_789.jpg', '35--Basketball/35_Basketball_playingbasketball_35_555.jpg', '30--Surgeons/30_Surgeons_Surgeons_30_107.jpg', '13--Interview/13_Interview_Interview_Sequences_13_973.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_442.jpg', '22--Picnic/22_Picnic_Picnic_22_564.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_577.jpg', '57--Angler/57_Angler_peoplefishing_57_20.jpg', '51--Dresses/51_Dresses_wearingdress_51_737.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_379.jpg', '51--Dresses/51_Dresses_wearingdress_51_77.jpg', '19--Couple/19_Couple_Couple_19_139.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_790.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_494.jpg', '18--Concerts/18_Concerts_Concerts_18_656.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_827.jpg', '19--Couple/19_Couple_Couple_19_319.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_540.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_10.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_163.jpg', '19--Couple/19_Couple_Couple_19_254.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_440.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_475.jpg', '45--Balloonist/45_Balloonist_Balloonist_45_857.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_359.jpg', '53--Raid/53_Raid_policeraid_53_107.jpg', '35--Basketball/35_Basketball_playingbasketball_35_113.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_972.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_408.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_403.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_259.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_214.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_243.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_633.jpg', '56--Voter/56_Voter_peoplevoting_56_641.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_736.jpg', '3--Riot/3_Riot_Riot_3_716.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_412.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_789.jpg', '30--Surgeons/30_Surgeons_Surgeons_30_819.jpg', '52--Photographers/52_Photographers_taketouristphotos_52_487.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_536.jpg', '48--Parachutist_Paratrooper/48_Parachutist_Paratrooper_Parachutist_Paratrooper_48_892.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_707.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_192.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_98.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_982.jpg', '39--Ice_Skating/39_Ice_Skating_iceskiing_39_541.jpg', '42--Car_Racing/42_Car_Racing_Car_Racing_42_857.jpg', '28--Sports_Fan/28_Sports_Fan_Sports_Fan_28_535.jpg', '41--Swimming/41_Swimming_Swimming_41_730.jpg', '56--Voter/56_Voter_peoplevoting_56_946.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_214.jpg', '30--Surgeons/30_Surgeons_Surgeons_30_43.jpg', '12--Group/12_Group_Group_12_Group_Group_12_578.jpg', '13--Interview/13_Interview_Interview_On_Location_13_636.jpg', '57--Angler/57_Angler_peoplefishing_57_251.jpg', '37--Soccer/37_Soccer_soccer_ball_37_60.jpg', '39--Ice_Skating/39_Ice_Skating_iceskiing_39_348.jpg', '46--Jockey/46_Jockey_Jockey_46_76.jpg', '10--People_Marching/10_People_Marching_People_Marching_10_People_Marching_People_Marching_10_524.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_641.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_176.jpg', '4--Dancing/4_Dancing_Dancing_4_854.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_364.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_245.jpg', '7--Cheering/7_Cheering_Cheering_7_687.jpg', '53--Raid/53_Raid_policeraid_53_827.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_133.jpg', '13--Interview/13_Interview_Interview_Sequences_13_864.jpg', '36--Football/36_Football_Football_36_80.jpg', '15--Stock_Market/15_Stock_Market_Stock_Market_15_241.jpg', '53--Raid/53_Raid_policeraid_53_686.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_747.jpg', '6--Funeral/6_Funeral_Funeral_6_870.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_266.jpg', '41--Swimming/41_Swimming_Swimming_41_238.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_396.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_702.jpg', '53--Raid/53_Raid_policeraid_53_14.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_703.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_353.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_708.jpg', '52--Photographers/52_Photographers_photographertakingphoto_52_479.jpg', '42--Car_Racing/42_Car_Racing_Nascar_42_442.jpg', '37--Soccer/37_Soccer_soccer_ball_37_28.jpg', '51--Dresses/51_Dresses_wearingdress_51_268.jpg', '61--Street_Battle/61_Street_Battle_streetfight_61_558.jpg', '51--Dresses/51_Dresses_wearingdress_51_377.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_629.jpg', '0--Parade/0_Parade_Parade_0_906.jpg', '47--Matador_Bullfighter/47_Matador_Bullfighter_matadorbullfighting_47_572.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_495.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_843.jpg', '61--Street_Battle/61_Street_Battle_streetfight_61_606.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_849.jpg', '58--Hockey/58_Hockey_icehockey_puck_58_653.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_267.jpg', '12--Group/12_Group_Team_Organized_Group_12_Group_Team_Organized_Group_12_274.jpg', '4--Dancing/4_Dancing_Dancing_4_769.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_222.jpg', '37--Soccer/37_Soccer_Soccer_37_651.jpg', '22--Picnic/22_Picnic_Picnic_22_152.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_415.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_45.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_902.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_648.jpg', '53--Raid/53_Raid_policeraid_53_489.jpg', '34--Baseball/34_Baseball_Baseball_34_143.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_901.jpg', '2--Demonstration/2_Demonstration_Protesters_2_542.jpg', '35--Basketball/35_Basketball_playingbasketball_35_252.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_734.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_76.jpg', '55--Sports_Coach_Trainer/55_Sports_Coach_Trainer_sportcoaching_55_196.jpg', '51--Dresses/51_Dresses_wearingdress_51_830.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_345.jpg', '16--Award_Ceremony/16_Award_Ceremony_Awards_Ceremony_16_85.jpg', '5--Car_Accident/5_Car_Accident_Car_Crash_5_574.jpg', '56--Voter/56_Voter_peoplevoting_56_118.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_birthdayparty_50_374.jpg', '12--Group/12_Group_Group_12_Group_Group_12_247.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_287.jpg', '4--Dancing/4_Dancing_Dancing_4_97.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_470.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_56.jpg', '7--Cheering/7_Cheering_Cheering_7_408.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_91.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_174.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_920.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_549.jpg', '18--Concerts/18_Concerts_Concerts_18_133.jpg', '3--Riot/3_Riot_Riot_3_101.jpg', '3--Riot/3_Riot_Riot_3_438.jpg', '57--Angler/57_Angler_peoplefishing_57_51.jpg', '37--Soccer/37_Soccer_soccer_ball_37_720.jpg', '37--Soccer/37_Soccer_Soccer_37_618.jpg', '52--Photographers/52_Photographers_photographertakingphoto_52_635.jpg', '11--Meeting/11_Meeting_Meeting_11_Meeting_Meeting_11_102.jpg', '42--Car_Racing/42_Car_Racing_Nascar_42_828.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_105.jpg', '4--Dancing/4_Dancing_Dancing_4_378.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_447.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_195.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_869.jpg', '13--Interview/13_Interview_Interview_Sequences_13_868.jpg', '4--Dancing/4_Dancing_Dancing_4_156.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_108.jpg', '2--Demonstration/2_Demonstration_Protesters_2_811.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_282.jpg', '2--Demonstration/2_Demonstration_Protesters_2_840.jpg', '57--Angler/57_Angler_peoplefishing_57_566.jpg', '45--Balloonist/45_Balloonist_Balloonist_45_518.jpg', '6--Funeral/6_Funeral_Funeral_6_745.jpg', '13--Interview/13_Interview_Interview_Sequences_13_209.jpg', '35--Basketball/35_Basketball_Basketball_35_754.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_85.jpg', '5--Car_Accident/5_Car_Accident_Car_Crash_5_38.jpg', '33--Running/33_Running_Running_33_517.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_135.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_358.jpg', '57--Angler/57_Angler_peoplefishing_57_1009.jpg', '0--Parade/0_Parade_marchingband_1_552.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_60.jpg', '12--Group/12_Group_Group_12_Group_Group_12_735.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_835.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_723.jpg', '9--Press_Conference/9_Press_Conference_Press_Conference_9_346.jpg', '32--Worker_Laborer/32_Worker_Laborer_Worker_Laborer_32_110.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_490.jpg', '37--Soccer/37_Soccer_soccer_ball_37_88.jpg', '14--Traffic/14_Traffic_Traffic_14_728.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_370.jpg', '21--Festival/21_Festival_Festival_21_513.jpg', '0--Parade/0_Parade_Parade_0_459.jpg', '12--Group/12_Group_Group_12_Group_Group_12_38.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_486.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_568.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_488.jpg', '23--Shoppers/23_Shoppers_Shoppers_23_328.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_310.jpg', '54--Rescue/54_Rescue_rescuepeople_54_191.jpg', '12--Group/12_Group_Group_12_Group_Group_12_315.jpg', '12--Group/12_Group_Group_12_Group_Group_12_101.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_411.jpg', '51--Dresses/51_Dresses_wearingdress_51_335.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_1005.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_64.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_544.jpg', '35--Basketball/35_Basketball_Basketball_35_529.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_477.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_163.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_640.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_915.jpg', '7--Cheering/7_Cheering_Cheering_7_404.jpg', '31--Waiter_Waitress/31_Waiter_Waitress_Waiter_Waitress_31_818.jpg', '36--Football/36_Football_Football_36_194.jpg', '54--Rescue/54_Rescue_rescuepeople_54_222.jpg', '6--Funeral/6_Funeral_Funeral_6_211.jpg', '50--Celebration_Or_Party/50_Celebration_Or_Party_houseparty_50_165.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_388.jpg', '21--Festival/21_Festival_Festival_21_601.jpg', '24--Soldier_Firing/24_Soldier_Firing_Soldier_Firing_24_890.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_451.jpg', '25--Soldier_Patrol/25_Soldier_Patrol_Soldier_Patrol_25_873.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_420.jpg', '12--Group/12_Group_Large_Group_12_Group_Large_Group_12_354.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_960.jpg', '0--Parade/0_Parade_Parade_0_519.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_256.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_242.jpg', '8--Election_Campain/8_Election_Campain_Election_Campaign_8_347.jpg', '2--Demonstration/2_Demonstration_Protesters_2_56.jpg', '41--Swimming/41_Swimming_Swimmer_41_704.jpg', '35--Basketball/35_Basketball_playingbasketball_35_219.jpg', '12--Group/12_Group_Group_12_Group_Group_12_62.jpg', '56--Voter/56_Voter_peoplevoting_56_819.jpg', '51--Dresses/51_Dresses_wearingdress_51_914.jpg', '3--Riot/3_Riot_Riot_3_184.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_66.jpg', '54--Rescue/54_Rescue_rescuepeople_54_102.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_417.jpg', '49--Greeting/49_Greeting_peoplegreeting_49_307.jpg', '2--Demonstration/2_Demonstration_Protesters_2_370.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_227.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_763.jpg', '17--Ceremony/17_Ceremony_Ceremony_17_588.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_391.jpg', '26--Soldier_Drilling/26_Soldier_Drilling_Soldiers_Drilling_26_307.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_239.jpg', '2--Demonstration/2_Demonstration_Demonstrators_2_41.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_698.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_777.jpg', '6--Funeral/6_Funeral_Funeral_6_733.jpg', '20--Family_Group/20_Family_Group_Family_Group_20_282.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_633.jpg', '61--Street_Battle/61_Street_Battle_streetfight_61_162.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_524.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_793.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_626.jpg', '2--Demonstration/2_Demonstration_Political_Rally_2_22.jpg', '19--Couple/19_Couple_Couple_19_836.jpg', '15--Stock_Market/15_Stock_Market_Stock_Market_15_569.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_642.jpg', '45--Balloonist/45_Balloonist_Balloonist_45_118.jpg', '10--People_Marching/10_People_Marching_People_Marching_2_822.jpg', '56--Voter/56_Voter_peoplevoting_56_228.jpg', '14--Traffic/14_Traffic_Traffic_14_644.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_343.jpg', '44--Aerobics/44_Aerobics_Aerobics_44_96.jpg', '39--Ice_Skating/39_Ice_Skating_iceskiing_39_138.jpg', '51--Dresses/51_Dresses_wearingdress_51_1035.jpg', '2--Demonstration/2_Demonstration_Protesters_2_179.jpg', '2--Demonstration/2_Demonstration_Protesters_2_91.jpg', '5--Car_Accident/5_Car_Accident_Accident_5_244.jpg', '2--Demonstration/2_Demonstration_Demonstration_Or_Protest_2_450.jpg', '35--Basketball/35_Basketball_basketballgame_ball_35_478.jpg', '39--Ice_Skating/39_Ice_Skating_Ice_Skating_39_504.jpg', '40--Gymnastics/40_Gymnastics_Gymnastics_40_1043.jpg', '29--Students_Schoolkids/29_Students_Schoolkids_Students_Schoolkids_29_489.jpg', '13--Interview/13_Interview_Interview_2_People_Visible_13_668.jpg', '21--Festival/21_Festival_Festival_21_562.jpg']
-val_labels = [[[648, 282, 52, 56]], [[553, 407, 36, 33]], [[430, 76, 142, 180]], [[458, 122, 150, 198]], [[397, 178, 173, 249]], [[262, 284, 626, 384]], [[547, 1228, 141, 105]], [[432, 432, 110, 157]], [[462, 102, 168, 228]], [[396, 36, 304, 406]], [[596, 144, 114, 142]], [[484, 337, 31, 37]], [[104, 816, 163, 245]], [[458, 120, 78, 102]], [[483, 305, 72, 109]], [[384, 132, 174, 315]], [[359, 169, 219, 292]], [[390, 158, 136, 134]], [[343, 172, 355, 481]], [[252, 84, 466, 256]], [[820, 128, 74, 96]], [[306, 258, 342, 453]], [[405, 258, 323, 426]], [[727, 84, 40, 69]], [[385, 95, 240, 296]], [[388, 65, 38, 37]], [[552, 92, 132, 150]], [[470, 130, 128, 150]], [[358, 36, 208, 310]], [[382, 10, 86, 112], [526, 50, 100, 102]], [[263, 372, 238, 205], [673, 314, 212, 222]], [[308, 124, 66, 100], [826, 166, 72, 98]], [[220, 529, 12, 20], [959, 434, 3, 3]], [[398, 220, 156, 188], [646, 334, 134, 198]], [[602, 301, 137, 183], [333, 446, 118, 145]], [[418, 289, 31, 34], [348, 246, 26, 29]], [[473, 178, 80, 101], [763, 383, 57, 78]], [[520, 390, 102, 135], [408, 375, 84, 105]], [[408, 243, 26, 24], [502, 85, 7, 11]], [[366, 160, 160, 212], [514, 130, 132, 204]], [[402, 16, 104, 126], [664, 148, 124, 130]], [[38, 374, 108, 136], [480, 272, 62, 84]], [[68, 14, 152, 184], [706, 462, 88, 124]], [[464, 95, 156, 209], [247, 247, 129, 171]], [[487, 581, 33, 55], [496, 613, 42, 37]], [[174, 697, 75, 123], [432, 661, 93, 99]], [[276, 84, 88, 146], [720, 94, 96, 158]], [[263, 134, 48, 70], [533, 460, 57, 72]], [[606, 42, 108, 160], [572, 68, 68, 150]], [[556, 78, 60, 66], [30, 180, 60, 62]], [[640, 206, 137, 250], [324, 176, 220, 351]], [[494, 356, 19, 28], [610, 352, 22, 31]], [[661, 334, 78, 99], [515, 681, 61, 79]], [[229, 92, 34, 51], [851, 318, 43, 23]], [[477, 169, 51, 55], [577, 125, 44, 47]], [[399, 202, 127, 169], [487, 547, 171, 239]], [[414, 178, 278, 370], [682, 122, 264, 358]], [[298, 84, 54, 90], [564, 212, 78, 80]], [[202, 96, 160, 218], [518, 80, 150, 188], [762, 66, 140, 196]], [[168, 44, 116, 154], [334, 84, 120, 134], [828, 58, 126, 172]], [[391, 114, 11, 13], [454, 101, 11, 16], [478, 107, 11, 12]], [[464, 240, 186, 214], [708, 378, 148, 174], [178, 494, 162, 132]], [[312, 266, 92, 102], [442, 268, 138, 148], [736, 94, 54, 112]], [[388, 128, 70, 136], [480, 186, 62, 96], [716, 142, 72, 128]], [[104, 66, 88, 144], [352, 138, 86, 120], [736, 80, 82, 130]], [[660, 156, 70, 98], [854, 202, 64, 102], [62, 184, 80, 102]], [[299, 162, 51, 68], [677, 463, 45, 53], [519, 595, 54, 76]], [[357, 21, 54, 77], [465, 23, 45, 63], [579, 39, 52, 73]], [[635, 144, 48, 59], [940, 380, 70, 65], [1, 343, 32, 51]], [[480, 130, 68, 72], [612, 164, 86, 80], [842, 50, 60, 84]], [[188, 265, 114, 161], [473, 134, 121, 151], [658, 191, 121, 175]], [[331, 135, 49, 74], [694, 180, 35, 53], [700, 45, 16, 18]], [[274, 152, 68, 110], [454, 148, 78, 106], [802, 230, 68, 94]], [[94, 100, 100, 150], [442, 218, 74, 118], [896, 66, 82, 126]], [[322, 248, 45, 56], [475, 231, 42, 51], [733, 201, 47, 56]], [[466, 287, 118, 109], [631, 188, 21, 18], [593, 142, 25, 24]], [[102, 124, 18, 23], [186, 136, 19, 24], [159, 142, 8, 8]], [[184, 453, 184, 208], [317, 144, 248, 285], [29, 421, 56, 69]], [[352, 380, 118, 52], [520, 342, 116, 62], [640, 410, 100, 56]], [[108, 40, 102, 166], [462, 126, 100, 162], [562, 50, 106, 156]], [[56, 200, 122, 146], [392, 140, 102, 158], [838, 164, 116, 140]], [[660, 268, 230, 238], [304, 96, 242, 272], [8, 2, 214, 250]], [[525, 250, 77, 85], [746, 372, 85, 65], [300, 258, 54, 53]], [[187, 42, 55, 75], [723, 109, 35, 48], [798, 114, 32, 42]], [[415, 300, 45, 53], [538, 224, 41, 52], [728, 239, 12, 15]], [[512, 231, 48, 78], [852, 483, 27, 37], [228, 447, 34, 44]], [[903, 631, 12, 15], [966, 634, 13, 15], [609, 291, 20, 37]], [[654, 230, 58, 84], [560, 236, 56, 82], [414, 250, 58, 82], [300, 246, 58, 90]], [[70, 145, 63, 70], [280, 205, 44, 49], [589, 171, 60, 63], [843, 141, 48, 57]], [[93, 266, 27, 34], [633, 280, 7, 9], [557, 259, 7, 6], [897, 304, 13, 9]], [[425, 346, 49, 108], [694, 359, 64, 120], [322, 193, 14, 17], [687, 176, 11, 18]], [[202, 210, 98, 140], [364, 252, 82, 116], [548, 308, 90, 110], [750, 202, 102, 156]], [[843, 119, 39, 51], [613, 153, 33, 46], [307, 127, 35, 52], [159, 118, 35, 44]], [[593, 280, 9, 14], [159, 257, 15, 31], [288, 354, 97, 100], [793, 522, 45, 38]], [[571, 86, 53, 79], [679, 379, 15, 29], [819, 214, 32, 41], [895, 442, 12, 14]], [[637, 174, 39, 52], [563, 180, 36, 48], [960, 195, 47, 69], [861, 186, 38, 54]], [[332, 74, 176, 240], [736, 62, 198, 320], [118, 42, 162, 206], [700, 48, 120, 188]], [[322, 133, 37, 41], [546, 103, 55, 54], [608, 141, 42, 49], [729, 241, 53, 42]], [[191, 232, 28, 31], [597, 177, 21, 30], [746, 297, 34, 37], [845, 197, 26, 44]], [[138, 346, 58, 84], [352, 364, 54, 74], [562, 380, 62, 80], [698, 326, 54, 94]], [[164, 226, 74, 112], [376, 306, 60, 82], [792, 140, 80, 112], [632, 294, 60, 86]], [[319, 122, 56, 71], [405, 190, 48, 60], [848, 277, 15, 17], [918, 250, 8, 13]], [[55, 150, 69, 78], [234, 145, 104, 159], [515, 240, 107, 139], [833, 87, 78, 107]], [[391, 256, 50, 57], [637, 264, 41, 51], [843, 243, 48, 67], [251, 94, 54, 61]], [[196, 488, 80, 116], [272, 92, 62, 74], [872, 72, 74, 110], [708, 534, 52, 52]], [[196, 123, 120, 157], [157, 811, 162, 235], [711, 817, 106, 137], [750, 1049, 90, 81]], [[754, 344, 60, 64], [658, 248, 72, 87], [518, 399, 54, 57], [275, 239, 75, 85]], [[358, 224, 69, 75], [544, 267, 45, 59], [767, 0, 56, 35], [504, 0, 41, 38]], [[236, 52, 66, 88], [344, 56, 72, 88], [480, 8, 74, 80], [634, 38, 68, 76]], [[249, 434, 14, 20], [296, 428, 13, 15], [378, 412, 10, 12], [547, 411, 12, 17]], [[212, 226, 58, 80], [384, 270, 54, 70], [542, 246, 60, 74], [718, 262, 72, 82]], [[210, 282, 108, 158], [300, 96, 106, 144], [368, 288, 108, 156], [504, 198, 156, 168]], [[882, 793, 21, 25], [789, 816, 18, 27], [723, 816, 21, 24], [637, 810, 21, 26]], [[369, 141, 36, 65], [381, 99, 40, 59], [616, 40, 43, 54], [726, 38, 34, 49]], [[711, 83, 54, 90], [154, 68, 82, 116], [333, 120, 51, 64], [68, 99, 30, 43]], [[167, 258, 68, 72], [143, 367, 210, 262], [705, 537, 31, 42], [614, 809, 25, 24]], [[200, 810, 112, 202], [298, 770, 150, 230], [306, 308, 390, 562], [790, 640, 166, 246], [2, 564, 104, 328]], [[715, 285, 50, 65], [826, 247, 48, 62], [552, 257, 49, 72], [353, 264, 47, 64], [322, 93, 38, 57]], [[240, 264, 96, 118], [352, 284, 66, 98], [440, 260, 104, 134], [562, 256, 72, 108], [644, 226, 126, 170]], [[205, 193, 35, 48], [77, 3, 31, 38], [233, 27, 25, 31], [647, 102, 24, 31], [890, 96, 23, 35]], [[24, 233, 81, 102], [127, 227, 59, 72], [173, 261, 40, 67], [344, 204, 69, 88], [625, 170, 53, 62]], [[141, 329, 180, 255], [520, 166, 161, 257], [371, 758, 174, 244], [753, 675, 196, 244], [581, 1088, 149, 188]], [[200, 70, 68, 110], [344, 118, 72, 98], [496, 74, 74, 108], [640, 120, 84, 102], [772, 82, 78, 108]], [[779, 338, 17, 26], [650, 355, 11, 18], [433, 301, 12, 21], [399, 324, 14, 19], [212, 222, 11, 12]], [[62, 154, 64, 82], [252, 182, 58, 74], [466, 100, 62, 76], [686, 150, 56, 74], [896, 26, 66, 82]], [[751, 308, 34, 36], [196, 362, 20, 25], [97, 396, 22, 34], [773, 396, 32, 40], [449, 319, 20, 31]], [[635, 191, 55, 74], [534, 207, 51, 69], [442, 185, 58, 70], [332, 179, 58, 72], [256, 197, 63, 76]], [[399, 123, 48, 63], [461, 237, 41, 52], [121, 66, 20, 29], [909, 150, 22, 36], [799, 177, 24, 32]], [[42, 32, 106, 200], [264, 148, 104, 150], [532, 276, 52, 62], [656, 202, 72, 118], [732, 194, 78, 166]], [[6, 217, 25, 39], [101, 224, 42, 42], [304, 218, 25, 40], [498, 307, 25, 41], [904, 294, 28, 41]], [[130, 238, 43, 61], [121, 374, 25, 61], [420, 226, 27, 35], [592, 142, 40, 47], [861, 364, 45, 93]], [[338, 909, 107, 118], [650, 926, 79, 101], [762, 138, 180, 256], [444, 118, 219, 253], [65, 127, 194, 250]], [[68, 270, 48, 65], [478, 261, 50, 74], [266, 246, 51, 77], [838, 258, 61, 78], [680, 273, 59, 75]], [[432, 113, 111, 144], [286, 281, 60, 77], [714, 308, 62, 84], [885, 353, 48, 58], [971, 377, 39, 47]], [[106, 498, 118, 146], [326, 484, 94, 136], [452, 474, 100, 142], [586, 284, 92, 146], [726, 178, 124, 182]], [[55, 133, 36, 58], [373, 172, 6, 11], [70, 227, 19, 33], [451, 271, 11, 25], [688, 160, 13, 30]], [[202, 304, 13, 16], [66, 141, 7, 9], [674, 211, 7, 11], [171, 71, 3, 5], [868, 58, 5, 7]], [[104, 246, 88, 84], [454, 110, 98, 104], [700, 156, 70, 88], [844, 234, 64, 78], [928, 182, 82, 100]], [[608, 179, 62, 81], [919, 97, 20, 22], [853, 89, 19, 24], [13, 107, 64, 76], [301, 148, 53, 70]], [[560, 3, 34, 39], [599, 64, 48, 60], [469, 77, 59, 67], [104, 19, 44, 55], [452, 12, 43, 57]], [[26, 438, 19, 32], [152, 435, 17, 28], [306, 396, 15, 30], [543, 389, 26, 33], [691, 562, 21, 28]], [[18, 372, 124, 116], [166, 314, 118, 150], [322, 380, 82, 108], [572, 220, 108, 174], [706, 200, 110, 182]], [[283, 173, 34, 42], [508, 165, 27, 39], [792, 48, 39, 55], [747, 163, 30, 37], [535, 196, 36, 35]], [[567, 32, 56, 76], [270, 13, 85, 67], [935, 40, 27, 33], [1003, 25, 20, 31], [796, 31, 26, 31]], [[317, 146, 176, 233], [581, 156, 187, 223], [957, 104, 59, 112], [9, 202, 53, 79], [895, 116, 46, 62]], [[165, 180, 54, 78], [271, 235, 51, 66], [409, 227, 49, 73], [523, 229, 52, 76], [681, 229, 58, 78], [823, 245, 62, 88]], [[779, 152, 30, 35], [730, 153, 30, 30], [413, 153, 12, 15], [353, 162, 17, 19], [131, 174, 13, 14], [165, 177, 11, 14]], [[289, 95, 83, 118], [100, 5, 67, 69], [363, 12, 78, 98], [568, 119, 58, 115], [456, 6, 29, 62], [820, 0, 42, 56]], [[136, 348, 70, 86], [270, 306, 54, 100], [434, 326, 56, 82], [590, 288, 70, 80], [736, 310, 58, 86], [872, 134, 102, 92]], [[788, 351, 44, 75], [805, 208, 50, 62], [529, 141, 34, 39], [403, 119, 33, 38], [120, 178, 49, 65], [133, 144, 35, 42]], [[159, 137, 9, 10], [849, 36, 17, 21], [685, 124, 7, 7], [208, 172, 7, 7], [388, 154, 8, 8], [346, 141, 5, 8]], [[245, 306, 19, 30], [298, 280, 22, 34], [380, 326, 18, 28], [342, 326, 16, 22], [535, 320, 17, 27], [475, 348, 17, 21]], [[396, 44, 82, 130], [632, 36, 80, 118], [882, 114, 84, 150], [460, 134, 114, 146], [480, 280, 184, 258], [8, 158, 286, 416]], [[582, 64, 43, 61], [710, 81, 55, 65], [803, 62, 46, 56], [911, 67, 49, 59], [439, 85, 96, 116], [220, 127, 60, 100]], [[212, 194, 72, 82], [346, 202, 66, 84], [428, 248, 58, 92], [592, 250, 60, 82], [624, 364, 66, 88], [780, 276, 60, 88]], [[202, 205, 19, 31], [226, 308, 20, 24], [309, 268, 23, 31], [367, 227, 30, 47], [636, 117, 45, 59], [971, 243, 24, 37]], [[111, 38, 30, 29], [292, 96, 23, 23], [373, 88, 20, 20], [394, 171, 33, 29], [199, 163, 26, 23], [631, 202, 41, 45]], [[689, 204, 111, 142], [564, 612, 51, 84], [15, 681, 37, 50], [604, 274, 80, 108], [464, 297, 87, 109], [237, 184, 91, 127]], [[799, 40, 47, 64], [444, 50, 61, 93], [246, 119, 39, 43], [560, 318, 18, 25], [679, 295, 16, 24], [346, 163, 29, 33]], [[642, 920, 30, 41], [460, 970, 32, 33], [151, 1013, 32, 45], [816, 1007, 37, 42], [657, 356, 58, 76], [321, 424, 65, 58]], [[561, 232, 47, 61], [726, 333, 55, 65], [891, 226, 57, 65], [792, 67, 46, 57], [426, 292, 50, 53], [167, 72, 54, 34]], [[836, 209, 42, 39], [653, 145, 17, 22], [451, 151, 18, 23], [216, 163, 15, 29], [273, 189, 8, 12], [295, 194, 6, 10]], [[819, 127, 56, 80], [699, 60, 60, 84], [478, 128, 49, 47], [137, 202, 38, 50], [74, 124, 47, 68], [264, 56, 27, 29]], [[858, 184, 30, 36], [711, 258, 21, 30], [516, 337, 16, 22], [581, 348, 15, 20], [448, 29, 45, 61], [121, 104, 44, 58]], [[10, 304, 74, 90], [242, 322, 74, 94], [488, 334, 84, 114], [822, 360, 72, 116], [662, 334, 68, 98], [460, 290, 66, 86]], [[746, 280, 18, 21], [798, 285, 15, 22], [219, 124, 24, 30], [259, 122, 26, 35], [363, 340, 68, 95], [123, 333, 81, 90]], [[446, 170, 114, 168], [600, 22, 108, 136], [132, 78, 116, 128], [8, 244, 92, 160], [756, 204, 108, 140], [876, 256, 102, 152]], [[175, 118, 17, 27], [297, 129, 10, 17], [329, 67, 53, 71], [459, 82, 53, 64], [534, 78, 51, 61], [757, 123, 20, 26]], [[58, 91, 55, 68], [113, 121, 48, 62], [276, 145, 41, 60], [333, 192, 28, 34], [486, 56, 68, 92], [601, 79, 48, 68]], [[150, 71, 88, 116], [315, 115, 53, 102], [675, 48, 81, 100], [779, 89, 69, 92], [620, 399, 20, 40], [529, 532, 28, 35]], [[394, 512, 166, 204], [196, 310, 304, 242], [522, 372, 252, 256], [170, 106, 268, 192], [438, 32, 208, 300], [642, 200, 198, 188]], [[920, 190, 68, 136], [687, 75, 96, 110], [392, 196, 73, 99], [325, 199, 55, 75], [288, 226, 22, 34], [549, 278, 34, 51]], [[27, 212, 20, 23], [159, 212, 17, 22], [221, 229, 17, 23], [317, 233, 36, 46], [667, 171, 131, 164], [138, 159, 9, 9]], [[287, 88, 34, 45], [404, 293, 11, 14], [578, 267, 12, 11], [735, 549, 7, 10], [930, 549, 10, 10], [779, 252, 10, 7]], [[450, 116, 68, 73], [529, 51, 65, 93], [904, 20, 61, 63], [150, 9, 43, 50], [669, 7, 42, 56], [215, 5, 40, 48], [306, 0, 21, 22]], [[254, 66, 66, 92], [104, 320, 72, 100], [282, 326, 106, 84], [456, 220, 62, 88], [684, 190, 70, 102], [820, 116, 72, 98], [726, 390, 76, 100]], [[542, 283, 7, 9], [883, 282, 9, 11], [819, 269, 9, 10], [755, 258, 20, 31], [289, 296, 30, 37], [511, 296, 17, 29], [613, 368, 30, 27]], [[321, 231, 43, 48], [164, 249, 53, 58], [41, 241, 53, 73], [423, 151, 37, 49], [617, 137, 39, 39], [645, 191, 48, 69], [930, 223, 43, 51]], [[914, 153, 26, 33], [792, 169, 19, 28], [678, 191, 18, 29], [320, 173, 13, 25], [256, 196, 14, 22], [184, 186, 19, 22], [100, 187, 21, 22]], [[327, 421, 12, 15], [265, 415, 11, 15], [250, 410, 12, 13], [422, 438, 10, 12], [501, 426, 10, 14], [471, 388, 11, 14], [501, 220, 11, 15]], [[496, 75, 87, 100], [412, 100, 72, 95], [431, 351, 66, 74], [983, 157, 37, 45], [798, 179, 22, 37], [627, 180, 52, 63], [77, 32, 36, 88]], [[416, 161, 122, 177], [101, 409, 113, 133], [540, 480, 99, 133], [790, 533, 106, 124], [544, 207, 99, 156], [767, 381, 90, 117], [131, 271, 87, 131]], [[84, 125, 74, 84], [426, 63, 54, 68], [288, 175, 28, 47], [426, 196, 8, 9], [517, 200, 27, 40], [685, 169, 25, 53], [865, 72, 80, 93]], [[498, 109, 19, 21], [385, 119, 21, 17], [196, 334, 9, 15], [924, 295, 14, 15], [485, 24, 14, 10], [627, 16, 12, 13], [238, 227, 14, 10]], [[58, 349, 32, 43], [215, 357, 30, 37], [385, 379, 11, 16], [1, 600, 28, 57], [598, 397, 32, 33], [328, 389, 6, 7], [349, 399, 5, 6]], [[0, 229, 23, 41], [777, 241, 9, 13], [831, 235, 11, 15], [894, 244, 10, 11], [913, 234, 11, 14], [942, 249, 8, 12], [318, 115, 67, 56]], [[89, 95, 71, 72], [337, 141, 45, 61], [429, 71, 38, 56], [505, 141, 45, 62], [556, 116, 36, 40], [694, 139, 51, 62], [872, 120, 60, 74]], [[524, 433, 96, 106], [154, 554, 71, 90], [936, 628, 78, 95], [36, 686, 67, 68], [910, 909, 14, 19], [811, 887, 20, 27], [0, 1011, 32, 62]], [[802, 185, 84, 65], [500, 290, 67, 93], [310, 284, 60, 70], [175, 318, 36, 57], [136, 260, 19, 25], [683, 368, 24, 27], [609, 384, 21, 33]], [[82, 108, 70, 74], [220, 148, 74, 74], [344, 152, 78, 102], [458, 98, 58, 72], [544, 164, 66, 84], [686, 150, 70, 86], [852, 138, 74, 84]], [[182, 375, 13, 19], [345, 368, 15, 18], [421, 376, 8, 14], [537, 372, 13, 17], [855, 392, 11, 15], [968, 385, 15, 17], [737, 384, 12, 13]], [[123, 387, 27, 37], [263, 394, 31, 42], [393, 387, 29, 38], [555, 314, 29, 39], [680, 392, 31, 40], [768, 390, 27, 35], [862, 400, 24, 39]], [[537, 248, 101, 138], [764, 228, 41, 48], [877, 261, 29, 29], [816, 387, 94, 104], [165, 199, 100, 153], [213, 300, 92, 113], [675, 247, 29, 47]], [[741, 275, 37, 40], [180, 272, 33, 34], [72, 336, 35, 54], [3, 287, 35, 39], [839, 308, 21, 45], [799, 121, 14, 16], [880, 122, 30, 36]], [[299, 83, 51, 71], [432, 143, 49, 93], [169, 0, 57, 40], [6, 202, 42, 86], [782, 49, 39, 77], [936, 134, 47, 90], [498, 185, 40, 84]], [[256, 238, 140, 199], [422, 223, 147, 224], [81, 634, 50, 72], [264, 659, 52, 68], [417, 671, 51, 62], [614, 708, 36, 57], [362, 981, 86, 52]], [[698, 55, 94, 100], [490, 333, 42, 39], [169, 242, 17, 20], [204, 139, 18, 14], [107, 193, 19, 15], [455, 12, 20, 26], [281, 152, 20, 10]], [[48, 3, 54, 56], [209, 2, 54, 75], [388, 14, 49, 72], [485, 18, 35, 62], [289, 262, 64, 78], [517, 227, 70, 75], [888, 186, 102, 129]], [[374, 174, 60, 54], [516, 168, 66, 64], [643, 30, 20, 28], [302, 9, 20, 29], [85, 31, 23, 31], [193, 21, 18, 22], [938, 9, 13, 17]], [[891, 75, 32, 53], [866, 138, 22, 26], [306, 79, 43, 61], [2, 106, 38, 47], [131, 145, 9, 9], [155, 203, 26, 34], [724, 84, 18, 39]], [[830, 0, 41, 24], [943, 32, 49, 67], [672, 66, 63, 84], [581, 7, 43, 49], [304, 145, 75, 97], [249, 12, 44, 50], [175, 59, 59, 71]], [[214, 118, 75, 102], [511, 111, 74, 92], [760, 114, 39, 90], [62, 166, 31, 44], [370, 152, 13, 30], [648, 212, 27, 38], [986, 368, 18, 23]], [[0, 438, 55, 161], [121, 656, 121, 150], [291, 636, 104, 169], [404, 740, 84, 139], [444, 780, 97, 167], [381, 908, 52, 94], [188, 815, 78, 155]], [[104, 77, 48, 59], [199, 78, 47, 51], [272, 69, 46, 59], [411, 88, 47, 51], [491, 89, 42, 51], [572, 76, 41, 51], [723, 68, 34, 45], [789, 69, 46, 49]], [[28, 260, 10, 16], [153, 458, 51, 54], [199, 419, 38, 48], [256, 365, 65, 85], [370, 386, 101, 121], [551, 398, 68, 93], [1003, 274, 17, 24], [924, 321, 46, 75]], [[105, 127, 23, 25], [180, 164, 10, 16], [281, 115, 5, 8], [251, 116, 5, 8], [361, 115, 30, 45], [702, 120, 14, 20], [910, 121, 13, 17], [922, 316, 26, 49]], [[169, 32, 39, 63], [229, 44, 119, 149], [471, 124, 63, 91], [383, 3, 56, 47], [511, 0, 142, 140], [981, 126, 43, 99], [0, 75, 39, 49], [92, 125, 25, 27]], [[547, 192, 11, 16], [577, 183, 14, 18], [621, 184, 14, 19], [315, 209, 33, 43], [478, 204, 27, 42], [334, 169, 23, 34], [321, 172, 16, 32], [130, 338, 76, 100]], [[100, 212, 64, 76], [284, 188, 54, 84], [468, 198, 60, 70], [654, 188, 58, 76], [834, 218, 60, 68], [758, 188, 90, 112], [542, 154, 114, 116], [358, 206, 114, 138]], [[448, 285, 50, 58], [666, 280, 33, 34], [810, 254, 58, 66], [787, 301, 16, 23], [932, 273, 40, 49], [81, 362, 18, 18], [979, 302, 15, 13], [670, 489, 40, 34]], [[818, 332, 45, 53], [720, 321, 45, 55], [613, 323, 36, 46], [520, 312, 38, 42], [401, 298, 36, 40], [331, 320, 31, 40], [274, 317, 32, 35], [196, 318, 33, 35]], [[166, 436, 51, 74], [350, 207, 51, 63], [321, 140, 52, 62], [479, 128, 50, 59], [520, 186, 49, 68], [661, 133, 56, 61], [730, 227, 51, 59], [829, 140, 52, 70]], [[365, 73, 58, 70], [527, 99, 65, 68], [675, 87, 65, 66], [763, 409, 50, 53], [600, 393, 54, 56], [481, 399, 57, 57], [321, 414, 57, 55], [222, 404, 50, 57]], [[405, 149, 114, 151], [601, 214, 42, 50], [579, 122, 47, 57], [727, 75, 32, 47], [334, 102, 50, 57], [212, 100, 58, 63], [279, 87, 36, 46], [395, 76, 26, 30]], [[323, 239, 201, 279], [546, 142, 56, 71], [674, 189, 57, 67], [789, 175, 50, 77], [828, 116, 53, 67], [863, 186, 53, 68], [960, 181, 60, 73], [437, 103, 31, 39]], [[850, 385, 27, 34], [793, 261, 20, 31], [634, 455, 30, 39], [583, 391, 33, 38], [638, 394, 20, 28], [520, 353, 24, 32], [634, 297, 11, 12], [925, 186, 10, 13]], [[147, 113, 66, 82], [256, 58, 75, 83], [385, 133, 64, 79], [558, 90, 67, 90], [703, 63, 73, 88], [851, 109, 69, 90], [695, 257, 71, 82], [637, 325, 60, 68]], [[266, 212, 29, 38], [376, 180, 28, 37], [440, 165, 28, 42], [504, 161, 32, 45], [564, 176, 29, 43], [639, 174, 30, 43], [697, 214, 29, 40], [766, 238, 28, 37]], [[876, 297, 65, 81], [875, 212, 47, 58], [848, 141, 40, 54], [832, 103, 35, 47], [726, 80, 39, 43], [432, 103, 50, 62], [870, 369, 54, 84], [54, 303, 43, 103]], [[894, 257, 33, 55], [842, 280, 35, 50], [571, 187, 97, 116], [556, 291, 23, 33], [341, 204, 72, 90], [191, 262, 64, 81], [98, 252, 43, 61], [21, 227, 40, 57]], [[929, 422, 27, 42], [873, 207, 22, 34], [627, 232, 22, 28], [649, 420, 26, 43], [378, 415, 25, 34], [413, 221, 22, 33], [133, 240, 22, 36], [93, 445, 24, 35]], [[417, 330, 241, 311], [60, 344, 17, 18], [200, 271, 19, 28], [299, 284, 21, 28], [263, 268, 13, 14], [768, 226, 10, 13], [799, 237, 8, 11], [914, 208, 11, 14]], [[305, 360, 38, 37], [474, 210, 30, 39], [584, 335, 39, 53], [720, 462, 23, 32], [131, 469, 22, 27], [523, 612, 30, 33], [395, 615, 29, 26], [470, 612, 21, 33]], [[436, 173, 77, 105], [183, 129, 111, 145], [144, 136, 30, 38], [721, 148, 68, 70], [576, 72, 73, 86], [644, 95, 57, 68], [774, 66, 59, 67], [665, 52, 45, 51]], [[69, 191, 11, 16], [405, 148, 10, 13], [456, 190, 12, 17], [421, 220, 15, 23], [453, 255, 18, 34], [555, 302, 19, 26], [826, 240, 18, 21], [991, 139, 9, 13]], [[509, 213, 41, 57], [677, 43, 47, 66], [845, 40, 51, 63], [325, 80, 52, 63], [217, 67, 40, 49], [173, 46, 39, 53], [124, 32, 31, 39], [51, 32, 35, 46]], [[18, 218, 35, 53], [129, 255, 27, 39], [216, 187, 50, 68], [375, 112, 54, 81], [493, 257, 27, 58], [543, 102, 54, 83], [739, 45, 70, 96], [889, 200, 31, 48]], [[191, 395, 15, 18], [298, 386, 48, 61], [403, 415, 50, 60], [491, 402, 14, 19], [474, 451, 42, 51], [522, 397, 39, 52], [589, 394, 42, 50], [490, 359, 9, 13]], [[33, 299, 68, 95], [198, 308, 77, 92], [240, 340, 104, 121], [376, 314, 74, 98], [447, 305, 115, 151], [627, 364, 83, 89], [778, 349, 71, 107], [888, 334, 65, 92]], [[16, 330, 18, 21], [304, 337, 26, 24], [403, 353, 26, 26], [512, 368, 16, 21], [645, 350, 16, 21], [768, 336, 16, 25], [871, 336, 17, 22], [933, 321, 24, 29]], [[750, 56, 63, 72], [477, 158, 20, 24], [389, 163, 21, 24], [302, 152, 25, 24], [215, 136, 27, 25], [129, 154, 26, 28], [42, 149, 27, 28], [61, 218, 48, 51]], [[52, 224, 138, 196], [204, 90, 114, 164], [514, 276, 160, 180], [226, 442, 184, 226], [726, 108, 124, 170], [850, 120, 122, 160], [834, 52, 66, 74], [804, 6, 70, 88]], [[43, 227, 12, 15], [63, 212, 13, 15], [102, 205, 17, 18], [218, 199, 15, 15], [231, 188, 24, 23], [332, 183, 11, 13], [408, 138, 37, 38], [488, 132, 44, 48], [747, 106, 28, 30]], [[363, 219, 81, 87], [390, 276, 60, 90], [411, 333, 63, 90], [598, 402, 81, 120], [760, 462, 90, 123], [474, 396, 78, 105], [321, 207, 63, 102], [168, 174, 57, 87], [246, 195, 60, 87]], [[23, 304, 79, 138], [211, 362, 73, 94], [283, 326, 54, 91], [353, 331, 56, 79], [444, 370, 39, 50], [656, 367, 44, 84], [819, 178, 205, 506], [795, 394, 82, 157], [778, 374, 44, 71]], [[706, 240, 63, 84], [543, 288, 50, 53], [633, 321, 26, 27], [591, 312, 26, 37], [585, 786, 37, 49], [206, 209, 144, 179], [111, 271, 52, 61], [396, 275, 60, 67], [0, 249, 62, 124]], [[86, 248, 34, 41], [123, 267, 42, 61], [176, 273, 71, 84], [415, 267, 115, 124], [595, 267, 26, 42], [667, 308, 46, 59], [750, 262, 37, 47], [722, 273, 29, 42], [859, 258, 98, 108]], [[0, 272, 65, 122], [140, 223, 67, 85], [173, 182, 88, 101], [326, 228, 66, 62], [447, 224, 85, 87], [572, 255, 73, 61], [710, 146, 48, 55], [813, 150, 65, 68], [957, 160, 50, 58]], [[17, 488, 28, 32], [164, 523, 54, 83], [271, 508, 30, 37], [432, 539, 38, 52], [800, 337, 53, 96], [745, 432, 9, 11], [775, 447, 17, 17], [12, 497, 6, 31], [977, 414, 9, 10]], [[211, 128, 33, 50], [750, 163, 29, 47], [21, 571, 16, 21], [178, 563, 9, 18], [476, 583, 9, 17], [852, 575, 13, 15], [921, 590, 14, 15], [959, 587, 10, 16], [1009, 575, 15, 18]], [[238, 394, 60, 68], [50, 404, 60, 72], [390, 398, 72, 90], [282, 552, 70, 70], [488, 468, 78, 86], [624, 328, 66, 82], [744, 438, 80, 84], [886, 276, 66, 82], [814, 194, 62, 82]], [[90, 348, 126, 184], [184, 344, 98, 176], [366, 248, 98, 158], [470, 194, 76, 132], [540, 228, 56, 100], [542, 90, 80, 130], [652, 90, 74, 126], [732, 94, 74, 120], [764, 42, 78, 112]], [[468, 968, 61, 69], [580, 871, 30, 37], [684, 927, 23, 34], [702, 1053, 66, 83], [815, 959, 68, 94], [924, 871, 68, 88], [246, 683, 96, 113], [829, 877, 22, 26], [759, 870, 28, 35]], [[189, 143, 47, 67], [257, 129, 45, 63], [341, 154, 49, 63], [399, 121, 42, 55], [467, 172, 46, 64], [556, 141, 42, 53], [650, 133, 53, 69], [744, 105, 45, 48], [802, 120, 52, 69]], [[647, 203, 38, 43], [754, 217, 39, 40], [860, 229, 40, 45], [116, 434, 42, 45], [293, 352, 37, 49], [386, 473, 38, 41], [482, 442, 38, 38], [279, 613, 44, 60], [603, 398, 42, 47]], [[823, 97, 66, 72], [630, 84, 30, 38], [452, 47, 52, 68], [126, 60, 65, 75], [445, 261, 37, 50], [448, 470, 25, 48], [819, 776, 132, 158], [548, 768, 80, 93], [97, 860, 105, 116]], [[608, 177, 20, 25], [691, 169, 20, 30], [729, 173, 19, 23], [753, 160, 23, 33], [916, 152, 35, 44], [968, 171, 31, 39], [68, 66, 45, 63], [212, 84, 32, 57], [127, 164, 29, 40]], [[21, 241, 34, 42], [264, 141, 6, 8], [406, 140, 7, 9], [459, 145, 5, 9], [462, 146, 6, 10], [490, 148, 6, 9], [608, 143, 5, 7], [672, 148, 6, 8], [956, 225, 14, 21]], [[408, 251, 30, 41], [547, 234, 12, 18], [971, 235, 6, 8], [989, 229, 7, 10], [1007, 231, 6, 9], [88, 232, 7, 10], [190, 233, 6, 9], [234, 241, 7, 9], [268, 242, 11, 14]], [[257, 246, 51, 75], [321, 245, 61, 80], [247, 190, 10, 14], [372, 165, 15, 17], [479, 165, 14, 21], [503, 283, 66, 87], [614, 201, 74, 105], [864, 177, 14, 19], [979, 165, 9, 10]], [[684, 279, 19, 24], [717, 281, 15, 26], [621, 273, 16, 23], [871, 309, 26, 28], [872, 282, 28, 32], [270, 237, 22, 29], [362, 251, 15, 25], [176, 273, 29, 36], [162, 339, 15, 23]], [[45, 267, 39, 53], [125, 240, 39, 55], [264, 239, 50, 58], [337, 215, 37, 79], [453, 233, 20, 22], [503, 213, 61, 72], [579, 240, 72, 77], [680, 222, 11, 12], [697, 183, 67, 71]], [[574, 462, 12, 14], [957, 466, 12, 14], [908, 450, 11, 14], [311, 491, 11, 14], [166, 500, 11, 15], [262, 495, 11, 15], [115, 498, 12, 15], [414, 644, 7, 11], [822, 610, 11, 12]], [[527, 157, 94, 125], [668, 187, 58, 69], [742, 218, 47, 58], [800, 246, 74, 98], [288, 111, 90, 118], [250, 152, 45, 64], [440, 200, 49, 58], [487, 215, 66, 89], [645, 236, 29, 37]], [[96, 178, 35, 49], [263, 215, 25, 40], [193, 342, 28, 42], [312, 355, 21, 38], [437, 235, 19, 35], [489, 275, 22, 28], [574, 219, 21, 37], [747, 198, 27, 38], [952, 174, 40, 46]], [[673, 265, 116, 141], [583, 360, 54, 63], [589, 393, 37, 50], [983, 222, 41, 100], [867, 656, 157, 112], [413, 617, 150, 147], [314, 104, 94, 107], [21, 109, 81, 122], [536, 388, 24, 35]], [[72, 391, 30, 31], [175, 378, 25, 30], [289, 364, 33, 29], [413, 354, 29, 33], [530, 360, 25, 41], [729, 360, 15, 21], [907, 343, 21, 24], [390, 210, 26, 34], [263, 350, 23, 32]], [[61, 404, 45, 57], [209, 352, 52, 62], [347, 219, 69, 79], [513, 347, 49, 63], [703, 81, 98, 73], [843, 242, 70, 69], [929, 358, 42, 60], [979, 400, 39, 53], [1, 438, 46, 33]], [[952, 497, 46, 62], [889, 515, 12, 14], [820, 494, 48, 59], [630, 437, 42, 60], [391, 377, 92, 160], [402, 181, 46, 51], [206, 195, 45, 50], [151, 538, 7, 9], [255, 93, 34, 36]], [[526, 471, 39, 51], [627, 272, 49, 28], [580, 275, 45, 45], [252, 46, 41, 49], [358, 77, 35, 39], [295, 121, 55, 40], [583, 46, 36, 36], [779, 92, 45, 37], [864, 43, 33, 34]], [[502, 151, 18, 22], [553, 163, 11, 13], [636, 154, 15, 17], [600, 171, 10, 13], [630, 174, 13, 16], [726, 172, 11, 13], [760, 151, 14, 17], [825, 155, 14, 14], [795, 152, 10, 17]], [[826, 138, 146, 159], [657, 148, 83, 102], [136, 200, 79, 91], [333, 302, 17, 18], [300, 305, 13, 16], [260, 311, 10, 13], [213, 292, 29, 35], [602, 279, 12, 18], [25, 281, 27, 38], [391, 308, 6, 11]], [[60, 356, 35, 51], [139, 339, 40, 38], [272, 301, 31, 41], [340, 284, 27, 28], [389, 271, 22, 32], [465, 267, 21, 27], [584, 300, 24, 20], [700, 265, 27, 35], [814, 301, 32, 28], [246, 376, 20, 51]], [[99, 88, 60, 91], [159, 287, 17, 21], [241, 281, 18, 25], [450, 466, 90, 41], [386, 301, 12, 14], [418, 298, 10, 13], [729, 323, 8, 10], [801, 328, 8, 11], [815, 320, 11, 12], [851, 331, 8, 11]], [[39, 440, 58, 40], [533, 253, 60, 83], [631, 269, 40, 46], [776, 239, 33, 59], [819, 127, 93, 129], [307, 189, 7, 8], [243, 410, 9, 10], [309, 331, 5, 7], [314, 330, 4, 6], [227, 110, 11, 13]], [[428, 67, 42, 45], [212, 36, 34, 43], [185, 159, 46, 48], [128, 237, 55, 33], [418, 317, 43, 40], [562, 312, 41, 36], [612, 359, 45, 44], [677, 543, 25, 57], [171, 450, 47, 47], [916, 61, 8, 10]], [[936, 169, 40, 52], [898, 18, 48, 58], [801, 117, 51, 51], [649, 181, 50, 50], [513, 99, 49, 56], [404, 85, 54, 51], [276, 77, 49, 62], [415, 10, 49, 51], [126, 88, 48, 52], [61, 135, 50, 57]], [[771, 339, 107, 138], [984, 255, 40, 91], [968, 164, 56, 64], [863, 152, 25, 39], [576, 150, 57, 84], [510, 164, 32, 38], [314, 167, 36, 44], [156, 175, 71, 95], [88, 180, 33, 46], [0, 398, 124, 152]], [[76, 390, 102, 107], [377, 326, 61, 91], [863, 404, 52, 81], [579, 465, 41, 78], [598, 483, 32, 53], [427, 441, 18, 34], [456, 380, 20, 27], [576, 413, 22, 32], [763, 362, 41, 54], [723, 424, 20, 39]], [[194, 110, 68, 104], [243, 86, 70, 113], [653, 50, 88, 149], [812, 86, 74, 135], [119, 455, 106, 155], [288, 466, 81, 151], [646, 432, 104, 185], [812, 473, 88, 169], [124, 839, 110, 178], [290, 860, 99, 169]], [[0, 254, 84, 251], [49, 342, 100, 129], [116, 373, 53, 80], [182, 439, 55, 65], [230, 373, 46, 57], [320, 438, 38, 38], [362, 458, 18, 23], [403, 435, 18, 24], [391, 470, 22, 29], [468, 464, 16, 20]], [[85, 282, 46, 55], [208, 274, 40, 53], [265, 344, 38, 44], [337, 319, 42, 48], [297, 296, 16, 24], [415, 277, 48, 53], [528, 297, 45, 47], [625, 292, 50, 56], [785, 299, 47, 58], [866, 252, 56, 60]], [[204, 285, 58, 93], [334, 208, 73, 93], [449, 265, 53, 76], [526, 203, 57, 68], [716, 88, 73, 88], [844, 138, 45, 62], [154, 24, 11, 18], [367, 337, 51, 85], [436, 218, 44, 55], [504, 99, 11, 14]], [[666, 222, 91, 101], [517, 152, 72, 98], [317, 76, 84, 106], [842, 260, 16, 20], [882, 264, 10, 14], [891, 251, 15, 20], [907, 244, 12, 22], [178, 134, 26, 48], [107, 150, 15, 19], [29, 150, 20, 24]], [[115, 76, 36, 67], [13, 207, 31, 51], [467, 402, 20, 43], [556, 272, 34, 46], [602, 304, 39, 53], [640, 245, 31, 44], [799, 328, 22, 34], [922, 323, 22, 34], [995, 493, 4, 5], [1004, 493, 3, 5]], [[588, 388, 19, 24], [644, 164, 14, 21], [523, 158, 15, 20], [408, 155, 15, 20], [654, 99, 14, 18], [473, 85, 14, 19], [336, 107, 16, 17], [806, 150, 15, 18], [440, 32, 14, 18], [353, 26, 12, 15]], [[943, 237, 14, 16], [927, 158, 13, 18], [898, 200, 20, 30], [797, 179, 13, 14], [761, 173, 18, 21], [664, 166, 8, 11], [590, 174, 29, 43], [688, 169, 10, 14], [387, 161, 22, 26], [288, 179, 38, 53]], [[944, 506, 38, 67], [962, 343, 41, 46], [847, 336, 12, 20], [616, 356, 17, 30], [611, 321, 16, 20], [515, 351, 19, 18], [441, 439, 39, 49], [453, 376, 20, 38], [429, 434, 27, 22], [189, 139, 19, 32]], [[20, 110, 39, 46], [90, 330, 45, 37], [283, 211, 78, 71], [388, 336, 35, 40], [553, 214, 95, 77], [787, 372, 55, 77], [257, 244, 14, 19], [2, 316, 26, 38], [49, 272, 28, 48], [29, 271, 17, 20]], [[855, 190, 94, 88], [576, 283, 95, 91], [411, 53, 82, 110], [170, 245, 112, 119], [734, 30, 19, 23], [598, 41, 12, 19], [643, 36, 16, 23], [263, 100, 19, 20], [288, 56, 17, 21], [325, 54, 20, 16]], [[644, 155, 30, 44], [847, 157, 23, 39], [983, 189, 13, 27], [850, 123, 29, 33], [514, 199, 33, 37], [158, 188, 30, 47], [102, 316, 12, 15], [81, 324, 13, 14], [52, 319, 11, 10], [4, 320, 12, 14]], [[959, 214, 38, 38], [784, 241, 22, 27], [815, 250, 14, 17], [731, 251, 14, 16], [666, 273, 10, 19], [632, 252, 13, 17], [594, 258, 12, 17], [535, 243, 15, 18], [406, 241, 15, 20], [341, 242, 27, 31]], [[462, 306, 22, 22], [493, 210, 21, 24], [626, 319, 46, 33], [649, 207, 20, 27], [706, 188, 25, 26], [798, 195, 23, 29], [568, 194, 20, 22], [841, 201, 27, 41], [920, 191, 18, 24], [998, 184, 26, 58]], [[188, 273, 45, 70], [292, 353, 19, 27], [319, 375, 16, 20], [583, 272, 35, 55], [729, 318, 17, 30], [812, 361, 8, 9], [861, 360, 9, 13], [1014, 361, 8, 14], [982, 343, 7, 12], [954, 339, 16, 21]], [[325, 470, 12, 12], [127, 411, 70, 102], [310, 281, 7, 10], [405, 261, 8, 11], [332, 261, 7, 9], [342, 248, 7, 8], [399, 393, 8, 10], [722, 320, 10, 12], [802, 337, 11, 14], [860, 276, 7, 11]], [[409, 158, 32, 46], [620, 231, 28, 33], [627, 133, 7, 10], [102, 179, 11, 17], [174, 213, 34, 46], [351, 151, 6, 8], [282, 132, 5, 8], [691, 134, 6, 11], [941, 155, 14, 19], [985, 170, 12, 11]], [[344, 430, 16, 19], [380, 431, 16, 22], [551, 483, 19, 22], [554, 421, 17, 19], [784, 412, 18, 18], [882, 430, 18, 20], [956, 413, 20, 22], [839, 464, 18, 12], [714, 540, 16, 22], [527, 317, 18, 10]], [[820, 63, 59, 69], [617, 146, 86, 101], [439, 240, 119, 135], [518, 17, 43, 50], [587, 2, 24, 42], [287, 22, 29, 43], [317, 12, 24, 37], [105, 20, 49, 58], [405, 70, 37, 43], [18, 178, 58, 107]], [[291, 470, 10, 14], [88, 686, 8, 16], [128, 716, 9, 15], [415, 474, 11, 14], [533, 472, 6, 14], [598, 471, 15, 20], [708, 477, 10, 19], [937, 454, 17, 23], [713, 628, 18, 21], [28, 554, 12, 14]], [[173, 158, 39, 44], [50, 243, 29, 30], [400, 272, 18, 32], [357, 292, 14, 19], [609, 121, 51, 57], [944, 244, 33, 47], [479, 340, 7, 6], [455, 362, 4, 5], [786, 373, 10, 13], [820, 387, 12, 12]]]
